% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_data.R
\name{arrow_enabled_object}
\alias{arrow_enabled_object}
\title{Determine whether arrow is able to serialize the given R object}
\usage{
arrow_enabled_object(object)
}
\arguments{
\item{object}{The object to be serialized}
}
\description{
If the given R object is not serializable by arrow due to some known
limitations of arrow, then return FALSE, otherwise return TRUE
}
\examples{
\dontrun{

df <- dplyr::tibble(x = seq(5))
arrow_enabled_object(df)
}

}
\keyword{internal}
