% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_data_frame.R
\name{as_spark_dataframe}
\alias{as_spark_dataframe}
\title{Convert objects of various types to Spark DataFrames}
\usage{
as_spark_dataframe(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional parameters.}
}
\description{
Convert objects of various types to Spark DataFrames
}

