% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_utils.R
\name{checkpoint_directory}
\alias{checkpoint_directory}
\alias{spark_set_checkpoint_dir}
\alias{spark_get_checkpoint_dir}
\title{Set/Get Spark checkpoint directory}
\usage{
spark_set_checkpoint_dir(sc, dir)

spark_get_checkpoint_dir(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{dir}{checkpoint directory, must be HDFS path of running on cluster}
}
\description{
Set/Get Spark checkpoint directory
}
