% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_table.R
\name{collect}
\alias{collect}
\alias{collect.tbl_spark}
\title{Collect results of dplyr expression}
\usage{
\method{collect}{tbl_spark}(x, ..., n = Inf, warn_incomplete = TRUE)
}
\arguments{
\item{x}{A \code{tbl_spark}.}

\item{n}{Limit on the number of records to copy}

\item{warn_incomplete}{Warn if all of the records are not copied}
}
\value{
An R data frame with classes \code{data.frame} and \code{tbl_df}.
}
\description{
Force computation of remote Spark DataFrame and copy the results back
to the client as an R data.frame
}

