% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{compile_package_jars}
\alias{compile_package_jars}
\title{Compile Scala sources into a Java Archive (jar)}
\usage{
compile_package_jars(package = rprojroot::find_package_root_file(),
  spark_versions = NULL, scalac = NULL, jar = NULL)
}
\arguments{
\item{package}{The path to an \R package.}

\item{spark_versions}{The Spark versions to build against. When \code{NULL},
builds against all Spark versions discovered with
\code{\link{spark_available_versions}}.}

\item{scalac}{The path to the \code{scalac} compiler to be used. When unset,
\code{scalac} will be discovered on the PATH.}

\item{jar}{The path to the \code{jar} Java archive tool to be used.
When unset, \code{jar} will be discovered on the PATH.}
}
\description{
Compile the \code{scala} source files contained within an \R package
into a Java Archive (\code{jar}) file that can be loaded and used within
a Spark environment.
}
\keyword{internal}

