% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{connection_is_open}
\alias{connection_is_open}
\title{Check whether the connection is open}
\usage{
connection_is_open(sc)
}
\arguments{
\item{sc}{\code{spark_connection}}
}
\description{
Check whether the connection is open
}
\keyword{internal}
