% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_shinyapp.R
\name{connection_spark_shinyapp}
\alias{connection_spark_shinyapp}
\title{A Shiny app that can be used to construct a \code{spark_connect} statement}
\usage{
connection_spark_shinyapp()
}
\description{
A Shiny app that can be used to construct a \code{spark_connect} statement
}
\keyword{internal}
