% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{copy_to}
\alias{copy_to}
\alias{copy_to.spark_connection}
\title{Copy a local R dataframe to Spark and provide a data source compatible with dplyr}
\usage{
\method{copy_to}{spark_connection}(dest, df, name = deparse(substitute(df)),
  ...)
}
\arguments{
\item{dest}{A Spark connection}

\item{df}{Local data frame to copy}

\item{name}{Name of the destination table}

\item{...}{Unused}
}
\description{
Copy a local R dataframe to Spark and provide a data source compatible with dplyr
}

