% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{copy_to.src_spark}
\alias{copy_to.src_spark}
\title{Copy a local R dataframe to Spark and provide a data source compatible with dplyr}
\usage{
\method{copy_to}{src_spark}(sc, df, name, cache = TRUE, repartition = 0,
  overwrite = TRUE)
}
\description{
Copy a local R dataframe to Spark and provide a data source compatible with dplyr
}
\keyword{internal}

