% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_spark_apply_bundle.R
\name{core_spark_apply_bundle}
\alias{core_spark_apply_bundle}
\title{Creates a bundle of dependencies required by \code{spark_apply()}}
\usage{
core_spark_apply_bundle()
}
\description{
Creates a bundle of dependencies required by \code{spark_apply()}
}
\keyword{internal}
