% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_spark_apply_bundle.R
\name{core_spark_apply_unbundle}
\alias{core_spark_apply_unbundle}
\title{Extracts a bundle of dependencies required by \code{spark_apply()}}
\usage{
core_spark_apply_unbundle(bundle_path)
}
\arguments{
\item{bundle_path}{Path to the bundle created using \code{core_spark_apply_bundle()}}
}
\description{
Extracts a bundle of dependencies required by \code{spark_apply()}
}
\keyword{internal}
