% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{db_data_type.src_spark}
\alias{db_data_type.src_spark}
\title{This operation is currently not supported in Spark}
\usage{
\method{db_data_type}{src_spark}(...)
}
\arguments{
\item{...}{Additional parameters}
}
\description{
This operation is currently not supported in Spark
}

