% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dbi_connection.R
\docType{methods}
\name{dbConnect,DBIConnectionSpark-method}
\alias{dbConnect,DBIConnectionSpark-method}
\title{Connect to Spark}
\usage{
\S4method{dbConnect}{DBIConnectionSpark}(drv, master = NULL, ...)
}
\arguments{
\item{drv}{\code{splyr::DBISpark()}}

\item{master}{Master location.}
}
\description{
Connect to Spark
}
\examples{
library(DBI)
con <- dbConnect(splyr::DBISpark())
dbDisconnect(con)
}

