% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dbi_connection.R
\docType{methods}
\name{dbDataType,DBIDriverSpark,ANY-method}
\alias{dbDataType,DBIDriverSpark,ANY-method}
\title{Determine database type for R vector.}
\usage{
\S4method{dbDataType}{DBIDriverSpark,ANY}(dbObj, obj)
}
\arguments{
\item{dbObj}{Spark driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector.
}
\keyword{internal}

