% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_driver.R
\docType{class}
\name{DBISpark}
\alias{DBISpark}
\alias{DBISparkDriver-class}
\alias{dbUnloadDriver,DBISparkDriver-method}
\title{Spark Driver}
\usage{
DBISpark()

\S4method{dbUnloadDriver}{DBISparkDriver}(drv, ...)
}
\description{
Spark Driver

DBISparkDriver and methods.
}
\examples{
library(DBI)
splyr::DBISpark()
}
\keyword{internal}

