% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dbi_driver.R
\docType{class}
\name{DBIDriverSpark-class}
\alias{DBIDriverSpark-class}
\alias{dbUnloadDriver,DBIDriverSpark-method}
\title{DBIDriverSpark and methods.}
\usage{
\S4method{dbUnloadDriver}{DBIDriverSpark}(drv, ...)
}
\description{
DBIDriverSpark and methods.
}
\keyword{internal}

