% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dbi_driver.R
\name{DBISpark}
\alias{DBISpark}
\title{Spark driver}
\usage{
DBISpark()
}
\description{
This driver never needs to be unloaded and hence \code{dbUnload()} is a
null-op.
}
\examples{
library(DBI)
splyr::DBISpark()
}

