% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_connection.R
\docType{class}
\name{DBISparkConnection-class}
\alias{DBISparkConnection-class}
\title{DBI Spark Connection.}
\description{
DBI Spark Connection.
}
\section{Slots}{

\describe{
\item{\code{scon}}{A Spark connection.}

\item{\code{api}}{Access to Spark API.}
}}
\keyword{internal}

