% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_driver.R
\docType{class}
\name{DBISparkDriver-class}
\alias{DBISparkDriver-class}
\title{DBI Spark Driver.}
\description{
DBI Spark Driver.
}
\section{Slots}{

\describe{
\item{\code{scon}}{A Spark connection.}
}}

