% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_result.R
\docType{class}
\name{DBISparkResult-class}
\alias{DBISparkResult-class}
\title{DBI Spark Result.}
\description{
DBI Spark Result.
}
\section{Slots}{

\describe{
\item{\code{sql}}{character.}

\item{\code{df}}{data.frame.}

\item{\code{lastFetch}}{numeric.}
}}
\keyword{internal}

