% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_connection.R
\docType{methods}
\name{dbSetProperty,DBISparkConnection,character,character-method}
\alias{dbSetProperty,DBISparkConnection,character,character-method}
\title{Sets a property for the connection}
\usage{
\S4method{dbSetProperty}{DBISparkConnection,character,character}(conn, property,
  value)
}
\arguments{
\item{conn}{DBI Spark connection}

\item{property}{The name of the property, for instance, "spark.sql.shuffle.partitions"}

\item{value}{The value to set the property to.}
}
\description{
Sets a property for the connection
}

