% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_connection.R
\name{dbSetProperty}
\alias{dbSetProperty}
\title{Sets a property for the connection}
\usage{
dbSetProperty(conn, property, value)
}
\arguments{
\item{conn}{DBI Spark connection}

\item{property}{The name of the property, for instance, "spark.sql.shuffle.partitions"}

\item{value}{The value to set the property to.}
}
\description{
Sets a property for the connection
}

