% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_options.R
\name{dplyr-spark-options-optimizeShuffleForCores}
\alias{dplyr-spark-options-optimizeShuffleForCores}
\title{The rspark.dplyr.optimizeShuffleForCores option optimizes the number of partitions
to use when shuffling data based on the number of local cores.
This option is only applicable to local installations. The default TRUE provides
automatic detection of cores. To avoid using this setting, set this value to FALSE.}
\description{
The rspark.dplyr.optimizeShuffleForCores option optimizes the number of partitions
to use when shuffling data based on the number of local cores.
This option is only applicable to local installations. The default TRUE provides
automatic detection of cores. To avoid using this setting, set this value to FALSE.
}
\examples{
getOption("rspark.dplyr.optimizeShuffleForCores", TRUE)
}

