% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_sql.R
\name{dplyr-sql}
\alias{dplyr-sql}
\alias{spark_sql_count_rows}
\title{Helper functions to support dplyr sql operations}
\usage{
spark_sql_count_rows(op, con)
}
\description{
Helper functions to support dplyr sql operations
}

