% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{find_scalac}
\alias{find_scalac}
\title{Discover the Scala Compiler}
\usage{
find_scalac(version, locations = NULL)
}
\arguments{
\item{version}{The \code{scala} version to search for. Versions
of the form \code{major.minor} will be matched against the
\code{scalac} installation with version \code{major.minor.patch};
if multiple compilers are discovered the most recent one will be
used.}

\item{locations}{Additional locations to scan. By default, the
directories \code{/opt/scala} and \code{/usr/local/scala} will
be scanned.}
}
\description{
Find the \code{scalac} compiler for a particular version of
\code{scala}, by scanning some common directories containing
\code{scala} installations.
}
