% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_vectors.R
\name{ft_pca}
\alias{ft_pca}
\alias{ml_pca}
\title{Feature Transformation -- PCA (Estimator)}
\usage{
ft_pca(
  x,
  input_col = NULL,
  output_col = NULL,
  k = NULL,
  uid = random_string("pca_"),
  ...
)

ml_pca(x, features = tbl_vars(x), k = length(features), pc_prefix = "PC", ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_col}{The name of the input column.}

\item{output_col}{The name of the output column.}

\item{k}{The number of principal components}

\item{uid}{A character string used to uniquely identify the feature transformer.}

\item{...}{Optional arguments; currently unused.}

\item{features}{The columns to use in the principal components
analysis. Defaults to all columns in \code{x}.}

\item{pc_prefix}{Length-one character vector used to prepend names of components.}
}
\value{
The object returned depends on the class of \code{x}. If it is a
\code{spark_connection}, the function returns a \code{ml_estimator} or a
\code{ml_estimator} object. If it is a \code{ml_pipeline}, it will return
a pipeline with the transformer or estimator appended to it. If a
\code{tbl_spark}, it will return a \code{tbl_spark} with the transformation
 applied to it.
}
\description{
PCA trains a model to project vectors to a lower dimensional space of the top k principal components.
}
\details{
In the case where \code{x} is a \code{tbl_spark}, the estimator
fits against \code{x} to obtain a transformer, returning a \code{tbl_spark}.

\code{ml_pca()} is a wrapper around \code{ft_pca()} that returns a
  \code{ml_model}.
}
\examples{
\dontrun{
library(dplyr)

sc <- spark_connect(master = "local")
iris_tbl <- sdf_copy_to(sc, iris, name = "iris_tbl", overwrite = TRUE)

iris_tbl \%>\%
  select(-Species) \%>\%
  ml_pca(k = 2)
}

}
\seealso{
Other feature transformers:
\code{\link[=ft_binarizer]{ft_binarizer()}},
\code{\link[=ft_bucketizer]{ft_bucketizer()}},
\code{\link[=ft_chisq_selector]{ft_chisq_selector()}},
\code{\link[=ft_count_vectorizer]{ft_count_vectorizer()}},
\code{\link[=ft_dct]{ft_dct()}},
\code{\link[=ft_elementwise_product]{ft_elementwise_product()}},
\code{\link[=ft_feature_hasher]{ft_feature_hasher()}},
\code{\link[=ft_hashing_tf]{ft_hashing_tf()}},
\code{\link[=ft_idf]{ft_idf()}},
\code{\link[=ft_imputer]{ft_imputer()}},
\code{\link[=ft_index_to_string]{ft_index_to_string()}},
\code{\link[=ft_interaction]{ft_interaction()}},
\code{\link{ft_lsh}},
\code{\link[=ft_max_abs_scaler]{ft_max_abs_scaler()}},
\code{\link[=ft_min_max_scaler]{ft_min_max_scaler()}},
\code{\link[=ft_ngram]{ft_ngram()}},
\code{\link[=ft_normalizer]{ft_normalizer()}},
\code{\link[=ft_one_hot_encoder]{ft_one_hot_encoder()}},
\code{\link[=ft_one_hot_encoder_estimator]{ft_one_hot_encoder_estimator()}},
\code{\link[=ft_polynomial_expansion]{ft_polynomial_expansion()}},
\code{\link[=ft_quantile_discretizer]{ft_quantile_discretizer()}},
\code{\link[=ft_r_formula]{ft_r_formula()}},
\code{\link[=ft_regex_tokenizer]{ft_regex_tokenizer()}},
\code{\link[=ft_robust_scaler]{ft_robust_scaler()}},
\code{\link[=ft_sql_transformer]{ft_sql_transformer()}},
\code{\link[=ft_standard_scaler]{ft_standard_scaler()}},
\code{\link[=ft_stop_words_remover]{ft_stop_words_remover()}},
\code{\link[=ft_string_indexer]{ft_string_indexer()}},
\code{\link[=ft_tokenizer]{ft_tokenizer()}},
\code{\link[=ft_vector_assembler]{ft_vector_assembler()}},
\code{\link[=ft_vector_indexer]{ft_vector_indexer()}},
\code{\link[=ft_vector_slicer]{ft_vector_slicer()}},
\code{\link[=ft_word2vec]{ft_word2vec()}}
}
\concept{feature transformers}
