% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ft_sql_transformer}
\alias{ft_sql_transformer}
\title{Feature Transformation -- SQLTransformer}
\usage{
ft_sql_transformer(x, sql, ...)
}
\arguments{
\item{x}{An object (usually a \code{spark_tbl}) coercable to a Spark DataFrame.}

\item{sql}{A SQL statement.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Transform a data set using SQL. Use the \code{__THIS__}
placeholder as a proxy for the active table.
}
