% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_standard_transformers.R
\name{ft_stop_words_remover}
\alias{ft_stop_words_remover}
\title{Feature Tranformation -- StopWordsRemover}
\usage{
ft_stop_words_remover(x, input_col, output_col, case_sensitive = FALSE,
  stop_words = ml_default_stop_words(spark_connection(x), "english"),
  uid = random_string("stop_words_remover_"), ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_col}{The name of the input column.}

\item{output_col}{The name of the output column.}

\item{case_sensitive}{Whether to do a case sensitive comparison over the stop words.}

\item{stop_words}{The words to be filtered out.}

\item{uid}{A character string used to uniquely identify the feature transformer.}

\item{...}{Optional arguments; currently unused.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection} When \code{x} is a \code{spark_connection}, the function returns a \code{ml_transformer},
  a \code{ml_estimator}, or one of their subclasses. The object contains a pointer to
  a Spark \code{Transformer} or \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline} When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the transformer or estimator appended to the pipline.

  \item \code{tbl_spark} When \code{x} is a \code{tbl_spark}, a transformer is constructed then
  immediately applied to the input \code{tbl_spark}, returning a \code{tbl_spark}
}
}
\description{
A feature transformer that filters out stop words from input.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

ml_default_stop_words

Other feature transformers: \code{\link{ft_one_hot_encoder}},
  \code{\link{ft_regex_tokenizer}},
  \code{\link{ft_tokenizer}}
}
