% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_invoke.R
\name{generic_call_interface}
\alias{generic_call_interface}
\title{Generic Call Interface}
\arguments{
\item{sc}{\code{spark_connection}}

\item{static}{Is this a static method call (including a constructor). If so
then the \code{object} parameter should be the name of a class (otherwise
it should be a spark_jobj instance).}

\item{object}{Object instance or name of class (for \code{static})}

\item{method}{Name of method}

\item{...}{Call parameters}
}
\description{
Generic Call Interface
}
