% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{get_spark_sql_catalog_implementation}
\alias{get_spark_sql_catalog_implementation}
\title{Retrieve the Spark connection's SQL catalog implementation property}
\usage{
get_spark_sql_catalog_implementation(sc)
}
\arguments{
\item{sc}{\code{spark_connection}}
}
\value{
spark.sql.catalogImplementation property from the connection's
 runtime configuration
}
\description{
Retrieve the Spark connection's SQL catalog implementation property
}
