% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{hive_context}
\alias{hive_context}
\title{Get the HiveContext associated with a connection}
\usage{
hive_context(sc)
}
\arguments{
\item{sc}{Connection to get HiveContext from}
}
\value{
Reference to HiveContext
}
\description{
Get the HiveContext \code{spark_jobj} associated with a
\code{spark_connection}
}

