% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_context_config.R
\name{hive_context_config}
\alias{hive_context_config}
\title{Runtime configuration interface for Hive}
\usage{
hive_context_config(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}
}
\description{
Retrieves the runtime configuration interface for Hive.
}
