% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_hof.R
\name{hof_exists}
\alias{hof_exists}
\title{dplyr wrapper for \code{exists(array<T>, function<T, Boolean>): Boolean}}
\usage{
hof_exists(x, dest_col, expr, pred, ...)
}
\arguments{
\item{x}{The Spark data frame to search}

\item{dest_col}{Column to store the search result}

\item{expr}{The array being searched (could be any SQL expression evaluating to an array)}

\item{pred}{A boolean predicate}

\item{...}{Additional params to dplyr::mutate}
}
\description{
dplyr wrapper for \code{exists(array<T>, function<T, Boolean>): Boolean}
}
