% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_hof.R
\name{hof_filter}
\alias{hof_filter}
\title{dplyr wrapper for \code{filter(array<T>, function<T, Boolean>): array<T>} in Spark SQL}
\usage{
hof_filter(x, func, expr = tail(colnames(x), 1), dest_col = expr, ...)
}
\arguments{
\item{x}{The Spark data frame to filter}

\item{func}{The filtering function}

\item{expr}{The array being filtered (could be any SQL expression evaluating to an array)}

\item{dest_col}{Column to store the filtered result}

\item{...}{Additional params to dplyr::mutate}
}
\description{
dplyr wrapper for \code{filter(array<T>, function<T, Boolean>): array<T>} in Spark SQL
}
