% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_hof.R
\name{hof_transform}
\alias{hof_transform}
\title{dplyr wrapper for \code{transform(array<T>, function<T, U>): array<U>} and
\code{transform(array<T>, function<T, Int, U>): array<U>} in Spark SQL}
\usage{
hof_transform(x, dest_col, expr, func, ...)
}
\arguments{
\item{x}{The Spark data frame to transform}

\item{dest_col}{Column to store the transformed result}

\item{expr}{The array being transformed (could be any SQL expression evaluating to an array)}

\item{func}{The transformation to apply}

\item{...}{Additional params to dplyr::mutate}
}
\description{
dplyr wrapper for \code{transform(array<T>, function<T, U>): array<U>} and
\code{transform(array<T>, function<T, Int, U>): array<U>} in Spark SQL
}
