% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_hof.R
\name{hof_zip_with}
\alias{hof_zip_with}
\title{dplyr wrapper for \code{zip_with(array<T>, array<U>, function<T, U, R>): array<R>}}
\usage{
hof_zip_with(x, dest_col, left, right, func, ...)
}
\arguments{
\item{x}{The Spark data frame to process}

\item{dest_col}{Column to store the query result}

\item{left}{Any expression evaluating to an array}

\item{right}{Any expression evaluating to an array}

\item{func}{Element-wise merge function to be applied}

\item{...}{Additional params to dplyr::mutate}
}
\description{
dplyr wrapper for \code{zip_with(array<T>, array<U>, function<T, U, R>): array<R>}
}
