% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_invoke.R
\name{invoke}
\alias{invoke}
\alias{invoke_new}
\alias{invoke_static}
\title{Execute a method on a remote Java object}
\usage{
invoke(jobj, method, ...)

invoke_static(sc, class, method, ...)

invoke_new(sc, class, ...)
}
\arguments{
\item{jobj}{Java object to execute method on.}

\item{method}{Name of method to execute.}

\item{...}{Unused (future expansion)}

\item{sc}{\code{spark_connection} to execute on.}

\item{class}{Class to execute static method on.}
}
\description{
Execute a method on a remote Java object
}

