% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_invoke.R
\name{j_invoke}
\alias{j_invoke}
\alias{j_invoke_static}
\alias{j_invoke_new}
\title{Invoke a Java function.}
\usage{
j_invoke(jobj, method, ...)

j_invoke_static(sc, class, method, ...)

j_invoke_new(sc, class, ...)
}
\arguments{
\item{jobj}{An \R object acting as a Java object reference (typically, a \code{spark_jobj}).}

\item{method}{The name of the method to be invoked.}

\item{...}{Optional arguments, currently unused.}

\item{sc}{A \code{spark_connection}.}

\item{class}{The name of the Java class whose methods should be invoked.}
}
\description{
Invoke a Java function and force return value of the call to be retrieved
as a Java object reference.
}
