% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{java_context}
\alias{java_context}
\title{Get the JavaSparkContext associated with a connection}
\usage{
java_context(sc)
}
\arguments{
\item{sc}{Connection to get SparkContext from}
}
\value{
Reference to SparkContext
}
\description{
Get the JavaSparkContext \code{spark_jobj} associated with a
\code{spark_connection}
}

