% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jfloat}
\alias{jfloat}
\title{Instantiate a Java float type.}
\usage{
jfloat(sc, x)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{x}{A numeric value in R.}
}
\description{
Instantiate a \code{java.lang.Float} object with the value specified.
NOTE: this method is useful when one has to invoke a Java/Scala method
requiring a float (instead of double) type for at least one of its
parameters.
}
\examples{
\donttest{
sc <- spark_connect(master = "spark://HOST:PORT")

jflt <- jfloat(sc, 1.23e-8)
# jflt is now a reference to a java.lang.Float object
}

}
