% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jfloat_array}
\alias{jfloat_array}
\title{Instantiate an Array[Float].}
\usage{
jfloat_array(sc, x)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{x}{A numeric vector in R.}
}
\description{
Instantiate an \code{Array[Float]} object with the value specified.
NOTE: this method is useful when one has to invoke a Java/Scala method
requiring an \code{Array[Float]} as one of its parameters.
}
\examples{
sc <- spark_connect(master = "spark://HOST:PORT")

jflt_arr <- jfloat_array(sc, c(-1.23e-8, 0, -1.23e-8))
# jflt_arr is now a reference an array of java.lang.Float

}
