% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_join.R
\name{join.tbl_spark}
\alias{join.tbl_spark}
\alias{inner_join.tbl_spark}
\alias{left_join.tbl_spark}
\alias{right_join.tbl_spark}
\alias{full_join.tbl_spark}
\title{Join Spark tbls.}
\usage{
\method{inner_join}{tbl_spark}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c("_x", "_y"),
  auto_index = FALSE,
  ...,
  sql_on = NULL
)

\method{left_join}{tbl_spark}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c("_x", "_y"),
  auto_index = FALSE,
  ...,
  sql_on = NULL
)

\method{right_join}{tbl_spark}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c("_x", "_y"),
  auto_index = FALSE,
  ...,
  sql_on = NULL
)

\method{full_join}{tbl_spark}(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c("_x", "_y"),
  auto_index = FALSE,
  ...,
  sql_on = NULL
)
}
\arguments{
\item{x}{tbls to join}

\item{y}{tbls to join}

\item{by}{a character vector of variables to join by.  If \code{NULL}, the
default, \code{*_join()} will do a natural join, using all variables with
common names across the two tables. A message lists the variables so
that you can check they're right (to suppress the message, simply
explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector.
For example, \code{by = c("a" = "b")} will match \code{x.a} to
\code{y.b}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into a
temporary table in same database as \code{x}. \code{*_join()} will automatically
run \code{ANALYZE} on the created table in the hope that this will make
you queries as efficient as possible by giving more data to the query
planner.

This allows you to join tables across srcs, but it's potentially expensive
operation so you must opt into it.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}

\item{auto_index}{if \code{copy} is \code{TRUE}, automatically create
indices for the variables in \code{by}. This may speed up the join if
there are matching indexes in \code{x}.}

\item{...}{other parameters passed onto methods, for instance, \code{na_matches}
to control how \code{NA} values are matched.  See \link{join.tbl_df} for more.}

\item{sql_on}{A custom join predicate as an SQL expression. The SQL
can refer to the \code{LHS} and \code{RHS} aliases to disambiguate
column names.}
}
\description{
These functions are wrappers around their `dplyr` equivalents that set
Spark SQL-compliant values for the `suffix` argument by replacing dots (`.`)
with underscores (`_`). See [join] for a description of the general purpose
of the functions.
}
