% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkster.R
\name{launch_ec2}
\alias{launch_ec2}
\title{Launch an EC2 Spark cluster}
\usage{
launch_ec2(access_key_id, secret_access_key, pem_path, instance_count = 1,
  version = "1.6.0", cluster_name = "sparkster",
  instance_type = "c3.4xlarge", region = "us-west-1", preview = FALSE)
}
\arguments{
\item{access_key_id}{EC2 access key id. Create a new access key from https://console.aws.amazon.com/iam/home?#security_credential}

\item{secret_access_key}{EC2 secret access key.}

\item{pem_path}{Identity file for ssh connections.}

\item{instance_count}{The total number of EC2 instances to be provisioned.}

\item{version}{The Spark version to use.}

\item{cluster_name}{Name used to identify cluster.}

\item{instance_type}{Type of EC2 instance. Tested with "m3.medium" and "c3.4xlarge".}

\item{region}{The EC2 region to host this cluster.}

\item{preview}{Print the EC2 command without executing anything?}
}
\description{
This function will launch a new an EC2 cluster.
}

