% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{list_sparklyr_jars}
\alias{list_sparklyr_jars}
\title{list all sparklyr-*.jar files that have been built}
\usage{
list_sparklyr_jars()
}
\description{
list all sparklyr-*.jar files that have been built
}
