% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livy_connection.R
\name{livy_config}
\alias{livy_config}
\title{Create a Spark Configuration for Livy}
\usage{
livy_config(config = spark_config(), username = NULL, password = NULL,
  csrf_header = TRUE)
}
\arguments{
\item{config}{Optional base configuration}

\item{username}{The username to use in the Authorization header}

\item{password}{The password to use in the Authorization header}

\item{csrf_header}{Required when the Livy server has CSRF protection enabled. Defaults to \code{TRUE}.}
}
\value{
Named list with configuration data
}
\description{
Create a Spark Configuration for Livy
}
\details{
Extends a Spark \code{"spark_config"} configuration with settings
for Livy. For instance, \code{"username"} and \code{"password"}
define the basic authentication settings for a Livy session.

When \code{"csrf_header"} is set to \code{TRUE}, the header \code{"X-Requested-By: sparklyr"}
is added to all HTTP requests.
}
