% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livy_connection.R
\name{livy_config_auth}
\alias{livy_config_auth}
\title{Create Spark Configuration with Basic Authentication}
\usage{
livy_config_auth(username, password, config = spark_config())
}
\arguments{
\item{username}{The username to use in the Authorization header}

\item{password}{The password to use in the Authorization header}

\item{config}{Optional base configuration}
}
\value{
Named list with configuration data
}
\description{
Create Spark Configuration with Basic Authentication
}
\details{
Extends a Spark \code{"spark_config"} configuration with
basic authentication for a Livy session.
}

