% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_data.R
\name{load_df}
\alias{load_df}
\title{Loads a dataframe and provides a data source compatible with dplyr}
\usage{
load_df(db, name, value, memory = TRUE, repartition = 0, overwrite = TRUE)
}
\arguments{
\item{db}{dplyr interface}

\item{name}{Name to reference the data source once it's loaded}

\item{value}{R data frame to load into Spark}

\item{memory}{Loads data into memory}

\item{repartition}{Total of partitions used to distribute table or 0 (default) to avoid partitioning}

\item{overwrite}{Overwrite the table with the given name when it exists}
}
\description{
Loads a dataframe and provides a data source compatible with dplyr
}

