% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_transformer_and_estimator.R
\name{ml-constructors}
\alias{ml-constructors}
\alias{ml_transformer}
\alias{ml_prediction_model}
\alias{ml_clustering_model}
\alias{ml_estimator}
\alias{ml_predictor}
\alias{ml_classifier}
\title{Constructors for Pipeline Stages}
\usage{
ml_transformer(jobj, ..., class = character())

ml_prediction_model(jobj, ..., class = character())

ml_clustering_model(jobj, ..., class = character())

ml_estimator(jobj, ..., class = character())

ml_predictor(jobj, ..., class = character())

ml_classifier(jobj, ..., class = character())
}
\arguments{
\item{jobj}{Pointer to the pipeline stage object.}

\item{...}{(Optional) additional attributes of the object.}

\item{class}{Name of class.}
}
\description{
Functions for developers writing extensions for Spark ML.
}
\keyword{internal}
