% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_transformer_and_estimator.R
\name{ml-estimators-transformers}
\alias{ml-estimators-transformers}
\alias{is_ml_transformer}
\alias{is_ml_estimator}
\alias{ml_fit}
\alias{ml_transform}
\alias{ml_fit_and_transform}
\title{Spark ML -- Transformers and Estimators}
\usage{
is_ml_transformer(x)

is_ml_estimator(x)

ml_fit(x, dataset, ...)

ml_transform(x, dataset, ...)

ml_fit_and_transform(x, dataset, ...)
}
\arguments{
\item{x}{An \code{ml_pipeline_stage} object.}

\item{dataset}{A \code{tbl_spark}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Methods for Spark ML Estimators and Transformers
}
