% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_param_utils.R
\name{ml-params}
\alias{ml-params}
\alias{ml_is_set}
\alias{ml_param_map}
\alias{ml_param}
\alias{ml_params}
\title{Spark ML -- ML Params}
\usage{
ml_is_set(x, param, ...)

ml_param_map(x, ...)

ml_param(x, param, allow_null = FALSE, ...)

ml_params(x, params = NULL, allow_null = FALSE, ...)
}
\arguments{
\item{x}{A Spark ML object, either a pipeline stage or an evaluator.}

\item{param}{The parameter to extract or set.}

\item{...}{Optional arguments; currently unused.}

\item{allow_null}{Whether to allow \code{NULL} results when extracting parameters. If \code{FALSE}, an error will be thrown if the specified parameter is not found. Defaults to \code{FALSE}.}

\item{params}{A vector of parameters to extract.}
}
\description{
Helper methods for working with parameters for ML objects.
}
