% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_transformer_and_estimator.R
\name{ml-pipelines}
\alias{ml-pipelines}
\alias{is_ml_transformer}
\alias{is_ml_estimator}
\title{Spark ML -- Transformers and Estimators}
\usage{
is_ml_transformer(x)

is_ml_estimator(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Transformer and Estimator objects.
}
