% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pipeline_utils.R
\name{ml_add_stage}
\alias{ml_add_stage}
\title{Add a Stage to a Pipeline}
\usage{
ml_add_stage(x, stage)
}
\arguments{
\item{x}{A pipeline or a pipeline stage.}

\item{stage}{A pipeline stage.}
}
\description{
Adds a stage to a pipeline.
}
\keyword{internal}
