% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_recommendation_als.R
\name{ml_als}
\alias{ml_als}
\title{Spark ML -- ALS}
\usage{
ml_als(x, rating_col = "rating", user_col = "user", item_col = "item",
  rank = 10L, reg_param = 0.1, implicit_prefs = FALSE, alpha = 1,
  nonnegative = FALSE, max_iter = 10L, num_user_blocks = 10L,
  num_item_blocks = 10L, checkpoint_interval = 10L,
  cold_start_strategy = "nan",
  intermediate_storage_level = "MEMORY_AND_DISK",
  final_storage_level = "MEMORY_AND_DISK", uid = random_string("als_"), ...)
}
\description{
Perform recommendation using Alternating Least Squares (ALS) matrix factorization.
}
