% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_als.R
\name{ml_als_tidiers}
\alias{ml_als_tidiers}
\alias{tidy.ml_als_model}
\alias{augment.ml_als_model}
\alias{glance.ml_als_model}
\title{Tidying methods for Spark ML ALS}
\usage{
\method{tidy}{ml_als_model}(x, ...)

\method{augment}{ml_als_model}(x, newdata = NULL, ...)

\method{glance}{ml_als_model}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}

\item{newdata}{a tbl_spark of new data to use for prediction.}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
