% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_apply_discrete_cosine_transform}
\alias{ml_apply_discrete_cosine_transform}
\title{Feature Transformation -- Discrete Cosine Transform (DCT)}
\usage{
ml_apply_discrete_cosine_transform(tbl, input_col, output_col,
  inverse = FALSE)
}
\arguments{
\item{tbl}{An object coercable to a Spark DataFrame.}

\item{input_col}{The name of the input column(s).}

\item{output_col}{The name of the output column.}

\item{inverse}{Perform inverse DCT?}
}
\description{
Transform a column in the time domain into another column in the frequency
domain.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ml_apply_binarizer}},
  \code{\link{ml_apply_bucketizer}},
  \code{\link{ml_apply_elementwise_product}},
  \code{\link{ml_apply_index_to_string}},
  \code{\link{ml_apply_quantile_discretizer}},
  \code{\link{ml_apply_sql_transformer}},
  \code{\link{ml_apply_string_indexer}},
  \code{\link{ml_apply_vector_assembler}}
}

