% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_apply_elementwise_product}
\alias{ml_apply_elementwise_product}
\title{Feature Transformation -- ElementwiseProduct}
\usage{
ml_apply_elementwise_product(df, input.col, output.col, scaling.col)
}
\arguments{
\item{df}{A Spark Dataset.}

\item{input.col}{The name of the input column(s).}

\item{output.col}{The name of the output column.}

\item{scaling.col}{The column used to scale \code{input.col}.}
}
\description{
Computes the element-wise product between two columns. Generally, this is
intended as a scaling transformation, where an input vector is scaled by
another vector, but this should apply for all element-wise product
transformations.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ml_apply_binarizer}},
  \code{\link{ml_apply_bucketizer}},
  \code{\link{ml_apply_discrete_cosine_transform}},
  \code{\link{ml_apply_index_to_string}},
  \code{\link{ml_apply_quantile_discretizer}},
  \code{\link{ml_apply_sql_transformer}},
  \code{\link{ml_apply_string_indexer}},
  \code{\link{ml_apply_vector_assembler}}
}

