% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_apply_quantile_discretizer}
\alias{ml_apply_quantile_discretizer}
\title{Feature Transformation -- QuantileDiscretizer}
\usage{
ml_apply_quantile_discretizer(tbl, input_col, output_col, n_buckets = 5)
}
\arguments{
\item{tbl}{An object coercable to a Spark DataFrame.}

\item{input_col}{The name of the input column(s).}

\item{output_col}{The name of the output column.}

\item{n_buckets}{The number of buckets to use.}
}
\description{
Takes a column with continuous features and outputs a column with binned
categorical features. The bin ranges are chosen by taking a sample of the
data and dividing it into roughly equal parts. The lower and upper bin bounds
will be -Infinity and +Infinity, covering all real values. This attempts to
find numBuckets partitions based on a sample of the given input data, but it
may find fewer depending on the data sample values.
}
\details{
Note that the result may be different every time you run it, since the sample
strategy behind it is non-deterministic.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ml_apply_binarizer}},
  \code{\link{ml_apply_bucketizer}},
  \code{\link{ml_apply_discrete_cosine_transform}},
  \code{\link{ml_apply_elementwise_product}},
  \code{\link{ml_apply_index_to_string}},
  \code{\link{ml_apply_sql_transformer}},
  \code{\link{ml_apply_string_indexer}},
  \code{\link{ml_apply_vector_assembler}}
}

