% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_apply_sql_transformer}
\alias{ml_apply_sql_transformer}
\title{Feature Transformation -- SQLTransformer}
\usage{
ml_apply_sql_transformer(df, input.col, output.col, sql)
}
\arguments{
\item{df}{A Spark Dataset.}

\item{input.col}{The name of the input column(s).}

\item{output.col}{The name of the output column.}

\item{sql}{A SQL statement.}
}
\description{
Transform a data set using SQL. Use the \code{__THIS__}
placeholder as a proxy for the active table.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ml_apply_binarizer}},
  \code{\link{ml_apply_bucketizer}},
  \code{\link{ml_apply_discrete_cosine_transform}},
  \code{\link{ml_apply_elementwise_product}},
  \code{\link{ml_apply_index_to_string}},
  \code{\link{ml_apply_quantile_discretizer}},
  \code{\link{ml_apply_string_indexer}},
  \code{\link{ml_apply_vector_assembler}}
}

