% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_apply_string_indexer}
\alias{ml_apply_string_indexer}
\title{Feature Transformation -- StringIndexer}
\usage{
ml_apply_string_indexer(x, input_col, output_col, params = NULL)
}
\arguments{
\item{x}{An object (usually a \code{spark_tbl}) coercable to a Spark DataFrame.}

\item{input_col}{The name of the input column(s).}

\item{output_col}{The name of the output column.}

\item{params}{An (optional) \R environment -- when available,
the index <-> label mapping generated by the string indexer
will be injected into this environment under the \code{labels}
key.}
}
\description{
Encode a column of labels into a column of label indices.
The indices are in [0, numLabels), ordered by label frequencies, with
the most frequent label assigned index 0. The transformation
can be reversed with \code{\link{ml_apply_index_to_string}}.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ml_apply_binarizer}},
  \code{\link{ml_apply_bucketizer}},
  \code{\link{ml_apply_discrete_cosine_transform}},
  \code{\link{ml_apply_elementwise_product}},
  \code{\link{ml_apply_index_to_string}},
  \code{\link{ml_apply_quantile_discretizer}},
  \code{\link{ml_apply_sql_transformer}},
  \code{\link{ml_apply_vector_assembler}}
}

