% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_apply_vector_assembler}
\alias{ml_apply_vector_assembler}
\title{Feature Transformation -- VectorAssembler}
\usage{
ml_apply_vector_assembler(x, input_col, output_col)
}
\arguments{
\item{x}{An object (usually a \code{spark_tbl}) coercable to a Spark DataFrame.}

\item{input_col}{The name of the input column(s).}

\item{output_col}{The name of the output column.}
}
\description{
Combine multiple vectors into a single row-vector; that is,
where each row element of the newly generated column is a
vector formed by concatenating each row element from the
specified input columns.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ml_apply_binarizer}},
  \code{\link{ml_apply_bucketizer}},
  \code{\link{ml_apply_discrete_cosine_transform}},
  \code{\link{ml_apply_elementwise_product}},
  \code{\link{ml_apply_index_to_string}},
  \code{\link{ml_apply_quantile_discretizer}},
  \code{\link{ml_apply_sql_transformer}},
  \code{\link{ml_apply_string_indexer}}
}

