% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification_evaluators.R
\name{ml_binary_classification_eval}
\alias{ml_binary_classification_eval}
\title{Spark ML - Binary Classification Evaluator}
\usage{
ml_binary_classification_eval(predicted_tbl_spark, label, score,
  metric = "areaUnderROC")
}
\arguments{
\item{predicted_tbl_spark}{The result of running sdf_predict}

\item{label}{Name of column string specifying which column contains the true, indexed labels (ie 0 / 1)}

\item{score}{Name of column contains the scored probability of a success (ie 1)}

\item{metric}{The classification metric -  one of: areaUnderRoc (default) or areaUnderPR}
}
\value{
area under the specified curve
}
\description{
See the Spark ML Documentation \href{https://spark.apache.org/docs/1.6.1/api/scala/index.html#org.apache.spark.ml.evaluation.BinaryClassificationEvaluator}{BinaryClassificationEvaluator}
}

