% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_evaluator.R
\name{ml_binary_classification_evaluator}
\alias{ml_binary_classification_evaluator}
\title{Spark ML - Binary Classification Evaluator}
\usage{
ml_binary_classification_evaluator(dataset, label_col, raw_prediction_col,
  metric_name = "areaUnderROC")
}
\arguments{
\item{dataset}{a \code{tbl_spark} containing a label column and a raw prediction column. This should be the output of \code{\link{sdf_predict}} or \code{\link{ml_predict}}.}

\item{label_col}{Name of column string specifying which column contains the true, indexed labels (ie 0 / 1)}

\item{raw_prediction_col}{Name of column contains the scored probability of a success (ie 1)}

\item{metric_name}{The classification metric - one of: areaUnderRoc (default) or areaUnderPR (not available in Spark 2.X)}
}
\value{
area under the specified curve
}
\description{
See the Spark ML Documentation \href{https://spark.apache.org/docs/1.6.1/api/scala/index.html#org.apache.spark.ml.evaluation.BinaryClassificationEvaluator}{BinaryClassificationEvaluator}
}
