% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering_bisecting_kmeans.R
\name{ml_bisecting_kmeans}
\alias{ml_bisecting_kmeans}
\title{Spark ML -- Bisecting K-Means Clustering}
\usage{
ml_bisecting_kmeans(x, k = 4L, max_iter = 20L, seed = NULL,
  min_divisible_cluster_size = 1, features_col = "features",
  prediction_col = "prediction",
  uid = random_string("bisecting_bisecting_kmeans_"), ...)
}
\description{
A bisecting k-means algorithm based on the paper "A comparison of document clustering techniques" by Steinbach, Karypis, and Kumar, with modification to fit Spark. The algorithm starts from a single cluster that contains all points. Iteratively it finds divisible clusters on the bottom level and bisects each of them using k-means, until there are k leaf clusters in total or no leaf clusters are divisible. The bisecting steps of clusters on the same level are grouped together to increase parallelism. If bisecting all divisible clusters on the bottom level would result more than k leaf clusters, larger clusters get higher priority.
}
