% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_constructor_utils.R
\name{ml_call_constructor}
\alias{ml_call_constructor}
\title{Wrap a Spark ML JVM object}
\usage{
ml_call_constructor(jobj)
}
\arguments{
\item{jobj}{The jobj for the pipeline stage.}
}
\description{
Identifies the associated sparklyr ML constructor for the JVM object by inspecting its
  class and performing a lookup. The lookup table is specified by the
  `sparkml/class_mapping.json` files of sparklyr and the loaded extensions.
}
\keyword{internal}
