% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classification_evaluators.R
\name{ml_classification_eval}
\alias{ml_classification_eval}
\title{Spark ML - Classification Evaluator}
\usage{
ml_classification_eval(predicted_tbl_spark, label, predicted_lbl,
  metric = "f1")
}
\arguments{
\item{predicted_tbl_spark}{A tbl_spark object that contains a columns with predicted labels}

\item{label}{Name of the column that contains the true, indexed label. Support for binary and multi-class labels, column should be of double type (use as.double)}

\item{predicted_lbl}{Name of the column that contains the predicted label NOT the scored probability. Support for binary and multi-class labels, column should be of double type (use as.double)}

\item{metric}{A classification metric, one of: f1 (default), precision, recall, weightedPrecision, weightedRecall, accuracy}
}
\value{
see \code{metric}
}
\description{
See the Spark ML Documentation \href{https://spark.apache.org/docs/1.6.1/api/scala/index.html#org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator}{MulticlassClassificationEvaluator}
}

