% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression_decision_tree_regressor.R
\name{ml_decision_tree_regressor}
\alias{ml_decision_tree_regressor}
\title{Spark ML -- Decision Trees}
\usage{
ml_decision_tree_regressor(x, features_col = "features",
  label_col = "label", prediction_col = "prediction", variance_col = NULL,
  checkpoint_interval = 10L, impurity = "variance", max_bins = 32L,
  max_depth = 5L, min_info_gain = 0, min_instances_per_node = 1L,
  seed = NULL, cache_node_ids = FALSE, max_memory_in_mb = 256L,
  uid = random_string("decision_tree_regressor_"), formula = NULL,
  response = NULL, features = NULL, ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{variance_col}{(Optional) Column name for the biased sample variance of prediction.}

\item{checkpoint_interval}{Set checkpoint interval (>= 1) or disable checkpoint (-1).
E.g. 10 means that the cache will get checkpointed every 10 iterations, defaults to 10.}

\item{impurity}{Criterion used for information gain calculation. Supported: "variance". (default = variance)}

\item{max_bins}{The maximum number of bins used for discretizing
continuous features and for choosing how to split on features at
each node. More bins give higher granularity.}

\item{max_depth}{Maximum depth of the tree (>= 0); that is, the maximum
number of nodes separating any leaves from the root of the tree.}

\item{min_info_gain}{Minimum information gain for a split to be considered
at a tree node. Should be >= 0, defaults to 0.}

\item{min_instances_per_node}{Minimum number of instances each child must
have after split.}

\item{seed}{Seed for random numbers.}

\item{cache_node_ids}{If \code{FALSE}, the algorithm will pass trees to executors to match instances with nodes.
If \code{TRUE}, the algorithm will cache node IDs for each instance. Caching can speed up training of deeper trees.
Defaults to \code{FALSE}.}

\item{max_memory_in_mb}{Maximum memory in MB allocated to histogram aggregation.
If too small, then 1 node will be split per iteration,
and its aggregates may exceed this size. Defaults to 256.}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{formula}{Used when \code{x} is a \code{tbl_spark}. R formula as a character string or a formula. This is used to transform the input dataframe before fitting, see \link{ft_r_formula} for details.}

\item{response}{(Deprecated) The name of the response column (as a length-one character
vector.)}

\item{features}{(Deprecated) The name of features (terms) to use for the model fit.}

\item{...}{Optional arguments; currently unused.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_predictor} object. The object contains a pointer to
  a Spark \code{Predictor} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the predictor appended to the pipline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, a predictor is constructed then
  immediately fit with the input \code{tbl_spark}, returning a prediction model.

  \item \code{tbl_spark}, with \code{formula}: specified When \code{formula}
    is specified, the input \code{tbl_spark} is first transformed using a
    \code{RFormula} transformer before being fit by
    the predictor. The object returned in this case is a \code{ml_model} which is a
    wrapper of a \code{ml_pipeline_model}.
}
}
\description{
Perform regression using decision trees.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-classification-regression.html} for
  more information on the set of supervised learning algorithms.
}
