% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_stop_words_remover.R
\name{ml_default_stop_words}
\alias{ml_default_stop_words}
\title{Default stop words}
\usage{
ml_default_stop_words(sc, language = c("english", "danish", "dutch",
  "finnish", "french", "german", "hungarian", "italian", "norwegian",
  "portuguese", "russian", "spanish", "swedish", "turkish"), ...)
}
\arguments{
\item{sc}{A \code{spark_connection}}

\item{language}{A character string.}

\item{...}{Optional arguments; currently unused.}
}
\value{
A list of stop words.
}
\description{
Loads the default stop words for the given language.
}
\details{
Supported languages: danish, dutch, english, finnish, french,
  german, hungarian, italian, norwegian, portuguese, russian, spanish,
  swedish, turkish. Defaults to English. See \url{http://anoncvs.postgresql.org/cvsweb.cgi/pgsql/src/backend/snowball/stopwords/}
  for more details
}
\seealso{
\code{\link{ft_stop_words_remover}}
}
