% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_helpers.R
\name{ml_feature_importances}
\alias{ml_feature_importances}
\alias{ml_tree_feature_importance}
\title{Spark ML - Feature Importance for Tree Models}
\usage{
ml_feature_importances(model, ...)

ml_tree_feature_importance(model, ...)
}
\arguments{
\item{model}{A decision tree-based model.}

\item{...}{Optional arguments; currently unused.}
}
\value{
For \code{ml_model}, a sorted data frame with feature labels and their relative importance.
  For \code{ml_prediction_model}, a vector of relative importances.
}
\description{
Spark ML - Feature Importance for Tree Models
}
