% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering_gaussian_mixture.R
\name{ml_gaussian_mixture}
\alias{ml_gaussian_mixture}
\title{Spark ML -- Gaussian Mixture clustering.}
\usage{
ml_gaussian_mixture(x, k = 2L, max_iter = 100L, tol = 0.01, seed = NULL,
  features_col = "features", prediction_col = "prediction",
  probability_col = "probability", uid = random_string("gaussian_mixture_"),
  ...)
}
\description{
This class performs expectation maximization for multivariate Gaussian Mixture Models (GMMs). A GMM represents a composite distribution of independent Gaussian distributions with associated "mixing" weights specifying each's contribution to the composite. Given a set of sample points, this class will maximize the log-likelihood for a mixture of k Gaussians, iterating until the log-likelihood changes by less than \code{tol}, or until it has reached the max number of iterations. While this process is generally guaranteed to converge, it is not guaranteed to find a global optimum.
}
