% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression_isotonic_regression.R
\name{ml_isotonic_regression}
\alias{ml_isotonic_regression}
\title{Spark ML -- Isotonic Regression}
\usage{
ml_isotonic_regression(x, feature_index = 0L, isotonic = TRUE,
  weight_col = NULL, features_col = "features", label_col = "label",
  prediction_col = "prediction",
  uid = random_string("isotonic_regression_"), ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{feature_index}{Index of the feature if \code{features_col} is a vector column (default: 0), no effect otherwise.}

\item{isotonic}{Whether the output sequence should be isotonic/increasing (true) or antitonic/decreasing (false). Default: true}

\item{weight_col}{The name of the column to use as weights for the model fit.}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments; currently unused.}

\item{formula}{Used when \code{x} is a \code{tbl_spark}. R formula as a character string or a formula. This is used to transform the input dataframe before fitting, see \link{ft_r_formula} for details.}

\item{features}{(Deprecated) The name of features (terms) to use for the model fit.}

\item{response}{(Deprecated) The name of the response column (as a length-one character
vector.)}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_predictor} object. The object contains a pointer to
  a Spark \code{Predictor} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the predictor appended to the pipline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, a predictor is constructed then
  immediately fit with the input \code{tbl_spark}, returning a prediction model.

  \item \code{tbl_spark}, with \code{formula}: specified When \code{formula}
    is specified, the input \code{tbl_spark} is first transformed using a
    \code{RFormula} transformer before being fit by
    the predictor. The object returned in this case is a \code{ml_model} which is a
    wrapper of a \code{ml_pipeline_model}.
}
}
\description{
Currently implemented using parallelized pool adjacent violators algorithm. Only univariate (single feature) algorithm supported.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-classification-regression.html} for
  more information on the set of supervised learning algorithms.

Other ml algorithms: \code{\link{ml_aft_survival_regression}},
  \code{\link{ml_decision_tree_classifier}},
  \code{\link{ml_gbt_classifier}},
  \code{\link{ml_generalized_linear_regression}},
  \code{\link{ml_linear_regression}},
  \code{\link{ml_linear_svc}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_multilayer_perceptron_classifier}},
  \code{\link{ml_naive_bayes}},
  \code{\link{ml_one_vs_rest}},
  \code{\link{ml_random_forest_classifier}}
}
