% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_kmeans.R
\name{ml_kmeans}
\alias{ml_kmeans}
\title{Computes kmeans from a dplyr source}
\usage{
ml_kmeans(x, centers, iter.max = 10)
}
\arguments{
\item{x}{A dplyr source.}

\item{centers}{Number of centers to compute.}

\item{iter.max}{Maximum number of iterations used to compute kmeans.}
}
\description{
Computes kmeans from a dplyr source
}

