% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model_kmeans.R
\name{ml_kmeans_cluster_eval}
\alias{ml_kmeans_cluster_eval}
\title{Evaluate a K-mean clustering}
\arguments{
\item{model}{A fitted K-means model returned by \code{ml_kmeans()}}

\item{dataset}{Dataset on which to calculate K-means cost}
}
\description{
Evaluate a K-mean clustering
}
