% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_lda.R
\name{ml_lda}
\alias{ml_lda}
\title{Spark ML -- Latent Dirichlet Allocation}
\usage{
ml_lda(x, features = tbl_vars(x), k = length(features), alpha = (50/k) +
  1, beta = 0.1 + 1, optimizer = "online", max.iterations = 20,
  ml.options = ml_options(), ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{features}{The name of features (terms) to use for the model fit.}

\item{k}{The number of topics to estimate.}

\item{alpha}{Concentration parameter for the prior placed on documents' distributions over topics. This is a singleton which is replicated to a vector of length \code{k} in fitting (as currently EM optimizer only supports symmetric distributions, so all values in the vector should be the same). For Expectation-Maximization optimizer values should be > 1.0.
By default \code{alpha = (50 / k) + 1}, where \code{50/k} is common in LDA libraries and +1 follows from Asuncion et al. (2009), who recommend a +1 adjustment for EM.}

\item{beta}{Concentration parameter for the prior placed on topics' distributions over terms. For Expectation-Maximization optimizer value should be > 1.0 and by default \code{beta = 0.1 + 1}, where 0.1 gives a small amount of smoothing and +1 follows Asuncion et al. (2009), who recommend a +1 adjustment for EM.}

\item{optimizer}{The optimizer, either \code{online} for Online Variational Bayes or \code{em} for Expectation-Maximization.}

\item{max.iterations}{Maximum number of iterations.}

\item{ml.options}{Optional arguments, used to affect the model generated. See
\code{\link{ml_options}} for more details.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Fit a Latent Dirichlet Allocation (LDA) model to a Spark DataFrame.
}
\note{
The topics' distributions over terms are called "beta" in the original LDA paper by Blei et al., but are called "phi" in many later papers such as Asuncion et al., 2009.

For terminology used in LDA model see \href{https://spark.apache.org/docs/latest/api/scala/index.html#org.apache.spark.ml.clustering.LDA}{Spark LDA documentation}.

Expectation-Maximization: Asuncion et al. \href{http://arxiv.org/pdf/1205.2662.pdf}{On Smoothing and Inference for Topic Models.} Uncertainty in Artificial Intelligence, 2009.
}
\examples{
\dontrun{
library(janeaustenr)
library(sparklyr)
library(dplyr)

sc <- spark_connect(master = "local")

austen_books <- austen_books()
books_tbl <- sdf_copy_to(sc, austen_books, overwrite = TRUE)
first_tbl <- books_tbl \%>\% filter(nchar(text) > 0) \%>\% head(100)

first_tbl \%>\%
  ft_tokenizer("text", "tokens") \%>\%
  ft_count_vectorizer("tokens", "features") \%>\%
  ml_lda("features", k = 4)
}
}
\references{
Original LDA paper (journal version): Blei, Ng, and Jordan. "Latent Dirichlet Allocation." JMLR, 2003.

Asuncion et al. (2009)
}
\seealso{
Other Spark ML routines: \code{\link{ml_kmeans}}
}
