% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_lda.R
\name{ml_lda_tidiers}
\alias{ml_lda_tidiers}
\alias{tidy.ml_lda_model}
\alias{augment.ml_lda_model}
\alias{glance.ml_lda_model}
\title{Tidying methods for Spark ML LDA models}
\usage{
\method{tidy}{ml_lda_model}(x, ...)

\method{augment}{ml_lda_model}(x, ...)

\method{glance}{ml_lda_model}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
