% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression.R
\name{ml_lm}
\alias{ml_lm}
\title{Linear regression from a dplyr source}
\usage{
ml_lm(x, response, features, intercept = TRUE)
}
\arguments{
\item{x}{A dplyr source.}

\item{response}{The prediction column}

\item{features}{List of columns to use as features}

\item{intercept}{TRUE to fit the intercept}
}
\description{
Linear regression from a dplyr source
}

