% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_logistic_regression.R
\name{ml_logistic_regression}
\alias{ml_logistic_regression}
\title{Logistic regression from a dplyr source}
\usage{
ml_logistic_regression(x, response, features, intercept = TRUE, alpha = 0,
  lambda = 0)
}
\arguments{
\item{x}{A dplyr source.}

\item{response}{The prediction column}

\item{features}{List of columns to use as features}

\item{intercept}{TRUE to fit the intercept}

\item{alpha}{The \emph{elastic net} mixing parameter.}

\item{lambda}{The \emph{regularization penalty}.}
}
\description{
Fit a logistic model using \code{spark.lm}.
}
\details{
See \url{https://spark.apache.org/docs/latest/ml-classification-regression.html}
for more information on how regression is implemented in Spark.
}

