% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_logistic_regression.R
\name{ml_logistic_regression_tidiers}
\alias{ml_logistic_regression_tidiers}
\alias{tidy.ml_model_logistic_regression}
\alias{augment.ml_model_logistic_regression}
\alias{augment._ml_model_logistic_regression}
\alias{glance.ml_model_logistic_regression}
\title{Tidying methods for Spark ML Logistic Regression}
\usage{
\method{tidy}{ml_model_logistic_regression}(x, ...)

\method{augment}{ml_model_logistic_regression}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_logistic_regression`}(x, new_data = NULL, ...)

\method{glance}{ml_model_logistic_regression}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}

\item{newdata}{a tbl_spark of new data to use for prediction.}

\item{new_data}{a tbl_spark of new data to use for prediction.}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
