% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{ml_model_data}
\alias{ml_model_data}
\title{Extracts data associated with a Spark ML model}
\usage{
ml_model_data(object)
}
\arguments{
\item{object}{a Spark ML model}
}
\value{
A tbl_spark
}
\description{
Extracts data associated with a Spark ML model
}
