% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_evaluator.R
\name{ml_multiclass_classification_evaluator}
\alias{ml_multiclass_classification_evaluator}
\title{Spark ML - Classification Evaluator}
\usage{
ml_multiclass_classification_evaluator(dataset, label_col, prediction_col,
  metric_name = "f1")
}
\arguments{
\item{dataset}{A tbl_spark object that contains a columns with predicted labels}

\item{label_col}{Name of the column that contains the true, indexed label. Support for binary and multi-class labels, column should be of double type (use as.double)}

\item{prediction_col}{Name of the column that contains the predicted label NOT the scored probability. Support for binary and multi-class labels, column should be of double type (use as.double)}

\item{metric_name}{A classification metric, for Spark 1.6: f1 (default), precision, recall, weightedPrecision, weightedRecall or accuracy;
for Spark 2.X: f1 (default), weightedPrecision, weightedRecall or accuracy}
}
\value{
see \code{metric}
}
\description{
See the Spark ML Documentation \href{https://spark.apache.org/docs/1.6.1/api/scala/index.html#org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator}{MulticlassClassificationEvaluator}
}
