% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_multilayer_perceptron.R
\name{ml_multilayer_perceptron}
\alias{ml_multilayer_perceptron}
\title{Spark ML -- Multilayer Perceptron}
\usage{
ml_multilayer_perceptron(x)
}
\arguments{
\item{x}{A \code{spark_tbl}.}
}
\description{
Creates and trains multilayer perceptron on a \code{spark_tbl}.
}
\seealso{
Other Spark ML routines: \code{\link{ml_kmeans}},
  \code{\link{ml_linear_regression}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_pca}}, \code{\link{ml_random_forest}}
}

