% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_multilayer_perceptron.R
\name{ml_multilayer_perceptron_tidiers}
\alias{ml_multilayer_perceptron_tidiers}
\alias{tidy.ml_model_multilayer_perceptron_classification}
\alias{augment.ml_model_multilayer_perceptron_classification}
\alias{glance.ml_model_multilayer_perceptron_classification}
\title{Tidying methods for Spark ML MLP}
\usage{
\method{tidy}{ml_model_multilayer_perceptron_classification}(x, ...)

\method{augment}{ml_model_multilayer_perceptron_classification}(x,
  newdata = NULL, ...)

\method{glance}{ml_model_multilayer_perceptron_classification}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}

\item{newdata}{a tbl_spark of new data to use for prediction.}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
