% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ml_mutate}
\alias{ml_mutate}
\alias{ml_mutate_}
\title{Mutate a Spark DataFrame}
\usage{
ml_mutate(.data, ...)

ml_mutate_(.data, ..., .dots)
}
\arguments{
\item{.data}{A \code{spark_tbl}.}

\item{...}{Named arguments, mapping new column names to the transformation to
be applied.}

\item{.dots}{A named list, mapping output names to transformations.}
}
\description{
Use Spark's \href{http://spark.apache.org/docs/latest/ml-features.html}{feature transformers}
to mutate a Spark DataFrame.
}

