% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_naive_bayes.R
\name{ml_naive_bayes}
\alias{ml_naive_bayes}
\title{Spark ML -- Naive-Bayes}
\usage{
ml_naive_bayes(x, response, features, lambda = 0, ...)
}
\arguments{
\item{x}{An object convertable to a Spark DataFrame (typically, a \code{tbl_spark}).}

\item{response}{The name of the response vector.}

\item{features}{The names of features (terms) included in the model.}

\item{lambda}{The (Laplace) smoothing parameter. Defaults to zero.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Perform regression or classification using naive bayes.
}
\seealso{
Other Spark ML routines: \code{\link{ml_decision_tree}},
  \code{\link{ml_gradient_boosted_trees}},
  \code{\link{ml_kmeans}}, \code{\link{ml_lda}},
  \code{\link{ml_linear_regression}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_multilayer_perceptron}},
  \code{\link{ml_one_vs_rest}}, \code{\link{ml_pca}},
  \code{\link{ml_random_forest}},
  \code{\link{ml_survival_regression}}
}

