% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_options.R
\name{ml_options}
\alias{ml_options}
\title{Provide Options for Spark.ML Routines}
\usage{
ml_options(feature.col = NULL, only.model = NULL, ...)
}
\arguments{
\item{feature.col}{The name to assign to the feature column.}

\item{only.model}{Boolean; should the Spark model object itself be returned
without fitting the actual model? Useful for \code{\link{ml_one_vs_rest}}.}

\item{...}{Optional arguments, reserved for future expansion.}
}
\description{
Provide Options for Spark.ML Routines
}

