% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{ml_partition}
\alias{ml_partition}
\title{Partition a Spark Dataframe}
\usage{
ml_partition(.data, ..., seed = sample(.Machine$integer.max, 1))
}
\arguments{
\item{.data}{Data}

\item{...}{Named parameters, mapping table names to weights.}

\item{seed}{Seed value for the partition}
}
\description{
Partition a Spark Dataframe
}

