% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pca.R
\name{ml_pca}
\alias{ml_pca}
\title{Perform Principal Components Analaysis using spark.ml}
\usage{
ml_pca(x, features = NULL)
}
\arguments{
\item{x}{A \code{tbl_spark}.}

\item{features}{The columns to use in the principal components
analysis. Defaults to all columns in \code{x}.}
}
\description{
Perform Principal Components Analaysis using spark.ml
}

