% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_pca.R
\name{ml_pca_tidiers}
\alias{ml_pca_tidiers}
\alias{tidy.ml_model_pca}
\alias{augment.ml_model_pca}
\alias{glance.ml_model_pca}
\title{Tidying methods for Spark ML Principal Component Analysis}
\usage{
\method{tidy}{ml_model_pca}(x, ...)

\method{augment}{ml_model_pca}(x, newdata = NULL, ...)

\method{glance}{ml_model_pca}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}

\item{newdata}{a tbl_spark of new data to use for prediction.}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
