% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_random_forest.R
\name{ml_random_forest}
\alias{ml_random_forest}
\title{Random Forests with Spark ML}
\usage{
ml_random_forest(x, response, features, max.bins = 32L, max.depth = 5L,
  num.trees = 20L)
}
\arguments{
\item{x}{A dplyr source.}

\item{response}{The response variable -- if the response is a factor or a logical
vector, then classification is performed (with the column treated as a label);
otherwise, regression is performed.}

\item{features}{List of columns to use as features.}

\item{max.bins}{Maximum number of bins.}

\item{max.depth}{Maximum depth.}

\item{num.trees}{Maximum number of trees.}
}
\description{
Random Forests with Spark ML
}

