% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_saveload.R
\name{ml_load}
\alias{ml_load}
\alias{ml_save}
\title{Save / Load a Spark ML Model Fit}
\usage{
ml_load(sc, file)

ml_save(model, file)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{file}{The filepath used for model save / load. Currently, only local filepaths
are supported.}

\item{model}{A \code{ml_model} fit.}
}
\description{
Save / load a \code{ml_model} fit.
}
\details{
These functions are currently experimental and not yet ready for production use.
}

