% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_pipeline_utils.R
\name{ml_stage}
\alias{ml_stage}
\alias{ml_stages}
\title{Spark ML -- Pipeline stage extraction}
\usage{
ml_stage(x, stage)

ml_stages(x, stages = NULL)
}
\arguments{
\item{x}{A \code{ml_pipeline} or a \code{ml_pipeline_model} object}

\item{stage}{The UID of a stage in the pipeline.}

\item{stages}{The UIDs of stages in the pipeline as a character vector.}
}
\value{
For \code{ml_stage()}: The stage specified.

For \code{ml_stages()}: A list of stages. If \code{stages} is not set, the function returns all stages of the pipeline in a list.
}
\description{
Extraction of stages from a Pipeline or PipelineModel object.
}
