% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_survival_regression.R
\name{ml_survival_regression}
\alias{ml_survival_regression}
\title{Spark ML -- Survival Regression}
\usage{
ml_survival_regression(x, response, features, intercept = TRUE,
  censor = "censor", ...)
}
\arguments{
\item{x}{An object convertable to a Spark DataFrame (typically, a \code{tbl_spark}).}

\item{response}{The name of the response vector.}

\item{features}{The names of features (terms) to use in the model.}

\item{intercept}{Fit the model with an intercept term?}

\item{censor}{The name of the vector that provides censoring information.
This should be a numeric vector, with 0 marking uncensored data, and
1 marking right-censored data.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Perform survival regression on a Spark DataFrame, using an Accelerated
failure time (AFT) model with potentially right-censored data.
}
\seealso{
Other Spark ML routines: \code{\link{ml_decision_tree}},
  \code{\link{ml_gradient_boosted_trees}},
  \code{\link{ml_kmeans}}, \code{\link{ml_lda}},
  \code{\link{ml_linear_regression}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_multilayer_perceptron}},
  \code{\link{ml_naive_bayes}},
  \code{\link{ml_one_vs_rest}}, \code{\link{ml_pca}},
  \code{\link{ml_random_forest}}
}

