% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model_print_methods.R
\name{ml_tree_feature_importance}
\alias{ml_tree_feature_importance}
\title{Spark ML - Feature Importance for Tree Models}
\usage{
ml_tree_feature_importance(sc, model)
}
\arguments{
\item{sc}{An active spark context}

\item{model}{An ml_model object of class decision trees (>1.5.0), random forest (>2.0.0), or GBT (>2.0.0)}
}
\value{
A sorted data frame with feature labels and their relative importance.
}
\description{
Spark ML - Feature Importance for Tree Models
}

