% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_tree_models.R
\name{ml_tree_tidiers}
\alias{ml_tree_tidiers}
\alias{tidy.ml_model_decision_tree_classification}
\alias{tidy.ml_model_decision_tree_regression}
\title{Tidying methods for Spark ML tree models}
\usage{
\method{tidy}{ml_model_decision_tree_classification}(x, ...)

\method{tidy}{ml_model_decision_tree_regression}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
