% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_unsupervised_models.R
\name{ml_unsupervised_tidiers}
\alias{ml_unsupervised_tidiers}
\alias{tidy.ml_model_kmeans}
\alias{augment.ml_model_kmeans}
\alias{glance.ml_model_kmeans}
\alias{tidy.ml_model_bisecting_kmeans}
\alias{augment.ml_model_bisecting_kmeans}
\alias{glance.ml_model_bisecting_kmeans}
\alias{tidy.ml_model_gaussian_mixture}
\alias{augment.ml_model_gaussian_mixture}
\alias{glance.ml_model_gaussian_mixture}
\title{Tidying methods for Spark ML unsupervised models}
\usage{
\method{tidy}{ml_model_kmeans}(x, ...)

\method{augment}{ml_model_kmeans}(x, newdata = NULL, ...)

\method{glance}{ml_model_kmeans}(x, ...)

\method{tidy}{ml_model_bisecting_kmeans}(x, ...)

\method{augment}{ml_model_bisecting_kmeans}(x, newdata = NULL, ...)

\method{glance}{ml_model_bisecting_kmeans}(x, ...)

\method{tidy}{ml_model_gaussian_mixture}(x, ...)

\method{augment}{ml_model_gaussian_mixture}(x, newdata = NULL, ...)

\method{glance}{ml_model_gaussian_mixture}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}

\item{newdata}{a tbl_spark of new data to use for prediction.}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
