% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_actions.R
\name{na.replace}
\alias{na.replace}
\title{Replace Missing Values in Objects}
\usage{
na.replace(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Arguments passed along to implementing methods.}
}
\description{
This S3 generic provides an interface for replacing
\code{\link{NA}} values within an object.
}
