% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_spark.R
\name{options-spark}
\alias{options-spark}
\title{The rspark.packages.default option, overrides the default Spark packages}
\description{
The rspark.packages.default option, overrides the default Spark packages

The rspark.install.dir option, defines the default Spark installation directory

The rspark.connection.allow_local_reconnect option, allows reconnect to
be enabled in local installations to help troubleshoot remote environments.

The rspark.connection.codegen option, sets the spark.sql.codegen.wholeStage option.

The spark.connection.codegen option, sets the spark.sql.codegen.wholeStage option.

The rspark.install.override.hive.site option, enables overriding the hive-site.xml config.
}
\examples{
getOption("rspark.packages.default", NULL)
getOption("rspark.install.dir", TRUE)
getOption("rspark.connection.allow_local_reconnect", FALSE)
getOption("rspark.connection.codegen", FALSE)
getOption("rspark.connection.codegen", FALSE)
getOption("rspark.install.override.hive.site", TRUE)
}

