% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{partition}
\alias{partition}
\title{Partition a Spark Dataframe}
\usage{
partition(.data, ..., seed = sample(.Machine$integer.max, 1))
}
\arguments{
\item{.data}{Data}

\item{...}{Named parameters, mapping table names to weights.}
}
\description{
Partition a Spark Dataframe
}

