% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_spark_jobj.R
\name{print.jobj}
\alias{print.jobj}
\title{Print a JVM object reference.}
\usage{
\method{print}{jobj}(x, ...)
}
\arguments{
\item{x}{The JVM object reference}

\item{...}{further arguments passed to or from other methods}
}
\description{
Prints information about a JVM object reference.
}
\details{
The underyling object id is printed; in addition, if the
associated connection is active; prints the class and
string representation of the underlying object.
}

