% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{print.src_spark}
\alias{print.src_spark}
\title{Prints information associated to the dplyr source}
\usage{
\method{print}{src_spark}(x, ...)
}
\arguments{
\item{x}{Reference to dplyr source}

\item{...}{Additional parameters}
}
\description{
Prints information associated to the dplyr source
}

