% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_jobj.R
\name{print_jobj}
\alias{print_jobj}
\title{Generic method for print jobj for a connection type}
\usage{
print_jobj(sc, jobj, ...)
}
\arguments{
\item{sc}{\code{spark_connection} (used for type dispatch)}

\item{jobj}{Object to print}
}
\description{
Generic method for print jobj for a connection type
}
\keyword{internal}
