% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{quote_sql_name}
\alias{quote_sql_name}
\title{Translate input character vector or symbol to a SQL identifier}
\usage{
quote_sql_name(x, con = NULL)
}
\description{
Calls dbplyr::translate_sql_ on the input character vector or symbol to obtain
the corresponding SQL identifier that is escaped and quoted properly
}
\keyword{internal}
