% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_string}
\alias{random_string}
\title{Random string generation}
\usage{
random_string(prefix = "table")
}
\arguments{
\item{prefix}{A length-one character vector.}
}
\description{
Generate a random string with a given prefix.
}
