% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_shiny.R
\name{reactiveSpark}
\alias{reactiveSpark}
\title{Reactive spark reader}
\usage{
reactiveSpark(x, intervalMillis = 1000,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame.}

\item{intervalMillis}{Approximate number of milliseconds to wait to retrieve
updated data frame. This can be a numeric value, or a function that returns
a numeric value.}

\item{session}{The user session to associate this file reader with, or NULL if
none. If non-null, the reader will automatically stop when the session ends.}
}
\description{
Given a spark object, returns a reactive data source for the contents
of the spark object. This function is most useful to read Spark streams.
}
