% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{spark_connection_is_open}
\alias{spark_log}
\alias{spark_web}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sparkapi}{\code{\link[sparkapi]{spark_connection_is_open}}, \code{\link[sparkapi]{spark_log}}, \code{\link[sparkapi]{spark_web}}}
}}
\keyword{internal}

