% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_extensions.R
\name{register_extension}
\alias{register_extension}
\alias{registered_extensions}
\title{Register a package that implements a Spark extension}
\usage{
register_extension(package)

registered_extensions()
}
\arguments{
\item{package}{Name of package to register}
}
\description{
Registering an extension package will result in the package being
automatically scanned for spark dependencies when a connection
to Spark is initiated (e.g. via \code{start_shell}).

Enumerate all registered extension packages
}
\note{
Extensions are typically registered when packages are
  loaded onto the search path (i.e. in the \code{.onLoad}
  function).
}

