% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_spark.R
\name{registerDoSpark}
\alias{registerDoSpark}
\title{explicitly registers a sparklyr parallel backend}
\usage{
registerDoSpark(spark_conn, ...)
}
\arguments{
\item{spark_conn}{spark connection to use}

\item{...}{additional options for sparklyr parallel backend
(currently only the only valid option is nocompile = {T, F})}
}
\value{
None
}
\description{
explicitly registers a sparklyr parallel backend
}
\examples{
sc <- spark_connect(master = "local")
registerDoSpark(sc, nocompile = FALSE)

}
