% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_data.R
\name{save_json}
\alias{save_json}
\title{Saves dplyr operation result as a JSON file}
\usage{
save_json(x, path)
}
\arguments{
\item{x}{A dplyr operation, for instance, `tbls(db, "flights")`}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports: "hdfs://" or "s3n://"}
}
\description{
Saves dplyr operation result as a JSON file
}

