% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_transformation_methods.R
\name{sdf-transform-methods}
\alias{sdf-transform-methods}
\alias{sdf_predict}
\alias{sdf_transform}
\alias{sdf_fit}
\alias{sdf_fit_and_transform}
\title{Spark ML -- Transform, fit, and predict methods (sdf_ interface)}
\usage{
sdf_predict(x, model, ...)

sdf_transform(x, transformer, ...)

sdf_fit(x, estimator, ...)

sdf_fit_and_transform(x, estimator, ...)
}
\arguments{
\item{x}{A \code{tbl_spark}.}

\item{model}{A \code{ml_transformer} or a \code{ml_model} object.}

\item{...}{Optional arguments passed to the corresponding \code{ml_} methods.}

\item{transformer}{A \code{ml_transformer} object.}

\item{estimator}{A \code{ml_estimator} object.}
}
\value{
\code{sdf_predict()}, \code{sdf_transform()}, and \code{sdf_fit_and_transform()} return a transformed dataframe whereas \code{sdf_fit()} returns a \code{ml_transformer}.
}
\description{
Methods for transformation, fit, and prediction. These are mirrors of the corresponding \link{ml-transform-methods}.
}
