% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_sequence.R
\name{sdf_along}
\alias{sdf_along}
\title{Create DataFrame for along Object}
\usage{
sdf_along(sc, along)
}
\arguments{
\item{sc}{The associated Spark connection.}

\item{along}{take the length from the length of this argument.}
}
\description{
Creates a DataFrame along the given object.
}
