% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_broadcast}
\alias{sdf_broadcast}
\title{Broadcast hint}
\usage{
sdf_broadcast(x)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}
}
\description{
Used to force broadcast hash joins.
}
