% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_checkpoint}
\alias{sdf_checkpoint}
\title{Checkpoint a Spark DataFrame}
\usage{
sdf_checkpoint(x, eager = TRUE)
}
\arguments{
\item{x}{an object coercible to a Spark DataFrame}

\item{eager}{whether to truncate the lineage of the DataFrame}
}
\description{
Checkpoint a Spark DataFrame
}
