% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_coalesce}
\alias{sdf_coalesce}
\title{Coalesces a Spark DataFrame}
\usage{
sdf_coalesce(x, partitions)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{partitions}{number of partitions}
}
\description{
Coalesces a Spark DataFrame
}
