% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_collect}
\alias{sdf_collect}
\title{Return a Spark DataFrame to R}
\usage{
sdf_collect(x)
}
\arguments{
\item{x}{a Spark DataFrame.}
}
\description{
Return a Spark DataFrame to R
}
\seealso{
Other Spark data frames: \code{\link{sdf_copy_to}},
  \code{\link{sdf_mutate}}, \code{\link{sdf_partition}},
  \code{\link{sdf_predict}}, \code{\link{sdf_register}},
  \code{\link{sdf_sample}}, \code{\link{sdf_sort}}
}

