% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_copy_to}
\alias{sdf_copy_to}
\alias{sdf_import}
\title{Copy an Object into Spark}
\usage{
sdf_copy_to(sc, x, ...)

sdf_import(x, sc, ...)
}
\arguments{
\item{sc}{The associated Spark connection.}

\item{x}{An \R object from which a Spark DataFrame can be generated.}

\item{...}{Optional arguments, passed to implementing methods.}
}
\description{
Copy an object into Spark, and return a Spark DataFrame.
}
\details{
\code{sdf_copy_to} is an S3 generic that, by default, dispatches to
\code{sdf_import}, itself also an S3 generic, but parameterized on the data
type rather than the connection type. If you'd like to implement
\code{sdf_copy_to} for a custom object type, you can accomplish this by
implementing the associated method on \code{sdf_import}.

A method for copying \R \code{data.frame}s to Spark is provided;
alternate methods can be provided by implementing the appropriate
}

