% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_stat.R
\name{sdf_crosstab}
\alias{sdf_crosstab}
\title{Cross Tabulation}
\usage{
sdf_crosstab(x, col1, col2)
}
\arguments{
\item{x}{A Spark DataFrame}

\item{col1}{The name of the first column. Distinct items will make the first item of each row.}

\item{col2}{The name of the second column. Distinct items will make the column names of the DataFrame.}
}
\value{
A DataFrame containing the contingency table.
}
\description{
Builds a contingency table at each combination of factor levels.
}
