% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{sdf_debug_string}
\alias{sdf_debug_string}
\title{Debug Info for Spark DataFrame}
\usage{
sdf_debug_string(x, print = TRUE)
}
\arguments{
\item{x}{An \R object wrapping, or containing, a Spark DataFrame.}

\item{print}{Print debug information?}
}
\description{
Prints plan of execution to generate \code{x}. This plan will, among other things, show the
number of partitions in parenthesis at the far left and indicate stages using indentation.
}
