% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_describe}
\alias{sdf_describe}
\title{Compute summary statistics for columns of a data frame}
\usage{
sdf_describe(x, cols = colnames(x))
}
\arguments{
\item{x}{An object coercible to a Spark DataFrame}

\item{cols}{Columns to compute statistics for, given as a character vector}
}
\description{
Compute summary statistics for columns of a data frame
}
