% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_dim.R
\name{sdf_dim}
\alias{sdf_dim}
\alias{sdf_dim}
\alias{sdf_nrow}
\alias{sdf_ncol}
\title{Support for Dimension Operations}
\usage{
sdf_dim(x)

sdf_nrow(x)

sdf_ncol(x)
}
\arguments{
\item{x}{An object (usually a \code{spark_tbl}).}
}
\description{
\code{sdf_dim()},  \code{sdf_nrow()} and \code{sdf_ncol()} provide similar
functionality to \code{dim()}, \code{nrow()} and \code{ncol()}.
}
