% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_drop_duplicates}
\alias{sdf_drop_duplicates}
\title{Remove duplicates from a Spark DataFrame}
\usage{
sdf_drop_duplicates(x, cols = NULL)
}
\arguments{
\item{x}{An object coercible to a Spark DataFrame}

\item{cols}{Subset of Columns to consider, given as a character vector}
}
\description{
Remove duplicates from a Spark DataFrame
}
