% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_wrapper.R
\name{sdf_extract_column}
\alias{sdf_extract_column}
\title{Extract Elements from a Vector Column}
\usage{
sdf_extract_column(x, column, ...)
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame (typically, a
\code{tbl_spark}).}

\item{column}{The column from which elements should
be extracted.}

\item{...}{Optional arguments, mapping column names
to the index at which the associated column element
can be extracted.}
}
\description{
Extract Elements from a Vector Column
}
