% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_from_avro}
\alias{sdf_from_avro}
\title{Convert column(s) from avro format}
\usage{
sdf_from_avro(x, cols = colnames(x))
}
\arguments{
\item{x}{An object coercible to a Spark DataFrame}

\item{cols}{Subset of Columns to convert from avro format}
}
\description{
Convert column(s) from avro format
}
