% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_streaming.R
\name{sdf_is_streaming}
\alias{sdf_is_streaming}
\title{Spark DataFrame is Streaming}
\usage{
sdf_is_streaming(x)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}
}
\description{
Is the given Spark DataFrame a streaming data?
}
