% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_last_index}
\alias{sdf_last_index}
\title{Returns the last index of a Spark DataFrame}
\usage{
sdf_last_index(x, id = "id")
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame (typically, a
\code{tbl_spark}).}

\item{id}{The name of the column to host the generated IDs.}
}
\description{
Returns the last index of a Spark DataFrame
}
