% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_num_partitions}
\alias{sdf_num_partitions}
\title{Gets number of partitions of a Spark DataFrame}
\usage{
sdf_num_partitions(x)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}
}
\description{
Gets number of partitions of a Spark DataFrame
}
