% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_wrapper.R
\name{sdf_read_column}
\alias{sdf_read_column}
\title{Read a Column from a Spark DataFrame}
\usage{
sdf_read_column(x, column)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{column}{The name of a column within \code{x}.}
}
\description{
Read a single column from a Spark DataFrame, and return
the contents of that column back to \R.
}
\details{
It is expected for this operation to preserve row order.
}
