% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{sdf_register}
\alias{sdf_register}
\title{Register a Spark DataFrame}
\usage{
sdf_register(x, name)
}
\arguments{
\item{x}{A Spark DataFrame.}

\item{name}{A name to assign this table.}
}
\description{
Registers a Spark DataFrame (giving it a table name for the
Spark SQL context), and returns a \code{tbl_spark}.
}

