% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_repartition}
\alias{sdf_repartition}
\title{Repartition a Spark DataFrame}
\usage{
sdf_repartition(x, partitions = NULL, partition_by = NULL)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{partitions}{number of partitions}

\item{partition_by}{vector of column names used for partitioning, only supported for Spark 2.0+}
}
\description{
Repartition a Spark DataFrame
}
