% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_regression_generalized_linear_regression.R,
%   R/ml_regression_linear_regression.R, R/ml_utils.R
\name{sdf_residuals.ml_model_generalized_linear_regression}
\alias{sdf_residuals.ml_model_generalized_linear_regression}
\alias{sdf_residuals.ml_model_linear_regression}
\alias{sdf_residuals}
\title{Model Residuals}
\usage{
\method{sdf_residuals}{ml_model_generalized_linear_regression}(object,
  type = c("deviance", "pearson", "working", "response"), ...)

\method{sdf_residuals}{ml_model_linear_regression}(object, ...)

sdf_residuals(object, ...)
}
\arguments{
\item{object}{Spark ML model object.}

\item{type}{type of residuals which should be returned.}

\item{...}{additional arguments}
}
\description{
This generic method returns a Spark DataFrame with model
residuals added as a column to the model training data.
}
