% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_stat.R
\name{sdf_rnorm}
\alias{sdf_rnorm}
\title{Generate random samples from the standard normal distribution}
\usage{
sdf_rnorm(sc, n, num_partitions = NULL, seed = NULL, output_col = "x")
}
\arguments{
\item{sc}{A Spark connection.}

\item{n}{Sample Size (default: 1000).}

\item{num_partitions}{Number of partitions in the resulting Spark dataframe
(default: default parallelism of the Spark cluster).}

\item{seed}{Random seed (default: a random long integer).}

\item{output_col}{Name of the output column containing sample values (default: "x").}
}
\description{
Generator method for creating a single-column Spark dataframes comprised of
i.i.d. samples from the standard normal distribution.
}
\seealso{
Other Spark statistical routines: 
\code{\link{sdf_rexp}()},
\code{\link{sdf_rgamma}()},
\code{\link{sdf_rlnorm}()},
\code{\link{sdf_rpois}()},
\code{\link{sdf_runif}()}
}
\concept{Spark statistical routines}
