% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{sdf_sample}
\alias{sdf_sample}
\title{Randomly Sample Rows from a Spark DataFrame}
\usage{
sdf_sample(x, fraction = 1, replacement = TRUE, seed = NULL)
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame.}

\item{fraction}{The fraction to sample.}

\item{replacement}{Boolean; sample with replacement?}

\item{seed}{An (optional) integer seed.}
}
\description{
Randomly Sample Rows from a Spark DataFrame
}

