% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_sequence.R
\name{sdf_seq}
\alias{sdf_seq}
\title{Create DataFrame for Range}
\usage{
sdf_seq(
  sc,
  from = 1L,
  to = 1L,
  by = 1L,
  repartition = NULL,
  type = c("integer", "integer64")
)
}
\arguments{
\item{sc}{The associated Spark connection.}

\item{from, to}{The start and end to use as a range}

\item{by}{The increment of the sequence.}

\item{repartition}{The number of partitions to use when distributing the
data across the Spark cluster. Defaults to the minimum number of partitions.}

\item{type}{The data type to use for the index, either \code{"integer"} or \code{"integer64"}.}
}
\description{
Creates a DataFrame for the given range
}
