% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_wrapper.R
\name{sdf_split_column}
\alias{sdf_split_column}
\title{Split a Vector Column into Scalar Columns}
\usage{
sdf_split_column(x, column, names = NULL)
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame (typically, a
\code{tbl_spark}).}

\item{column}{The name of a (vector-typed) column.}

\item{names}{The names to be assigned to the generated columns.}
}
\description{
Given a vector column in a Spark DataFrame,
split that into \code{n} separate columns, each
column made up of the different elements in the
column \code{column}.
}
