% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_to_avro}
\alias{sdf_to_avro}
\title{Convert column(s) to avro format}
\usage{
sdf_to_avro(x, cols = colnames(x))
}
\arguments{
\item{x}{An object coercible to a Spark DataFrame}

\item{cols}{Subset of Columns to convert into avro format}
}
\description{
Convert column(s) to avro format
}
