% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_weighted_quantile}
\alias{sdf_weighted_quantile}
\title{Compute (Approximate) Weighted Quantiles with a Spark DataFrame}
\usage{
sdf_weighted_quantile(
  x,
  column,
  weightColumn,
  probabilities = c(0, 0.25, 0.5, 0.75, 1),
  max.error = 0.01
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{column}{The column(s) for which quantiles should be computed.}

\item{weightColumn}{The column containing relative weight for each data point.}

\item{probabilities}{A numeric vector of probabilities, for
which quantiles should be computed.}

\item{max.error}{The maximal possible difference between the actual
percentile of a result and its expected percentile (e.g., if `max.error` is
0.01 and `probabilities` is 0.95, then any value between the 94th and 96th
weighted percentile will be considered an acceptable approximation).}
}
\description{
Given a value column within a Spark DataFrame and a weight column containing
relative weight for each value, compute approximate weighted quantiles.
}
