% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_with_sequential_id}
\alias{sdf_with_sequential_id}
\title{Add a Sequential ID Column to a Spark DataFrame}
\usage{
sdf_with_sequential_id(x, id = "id", from = 1L)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{id}{The name of the column to host the generated IDs.}

\item{from}{The starting value of the id column}
}
\description{
Add a sequential ID column to a Spark DataFrame. The Spark
\code{zipWithIndex} function is used to produce these. This differs from
\code{sdf_with_unique_id} in that the IDs generated are independent of
partitioning.
}
