% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_with_unique_id}
\alias{sdf_with_unique_id}
\title{Add a Unique ID Column to a Spark DataFrame}
\usage{
sdf_with_unique_id(x, id = "id")
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{id}{The name of the column to host the generated IDs.}
}
\description{
Add a unique ID column to a Spark DataFrame. The Spark
\code{monotonicallyIncreasingId} function is used to produce these and is
guaranteed to produce unique, monotonically increasing ids; however, there
is no guarantee that these IDs will be sequential. The table is persisted
immediately after the column is generated, to ensure that the column is
stable -- otherwise, it can differ across new computations.
}
