% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_local.R
\name{setup_local}
\alias{setup_local}
\title{Prepares the spark environment, downloading the spark binaries
to a common directory if needed.}
\usage{
setup_local(version = "1.6.0")
}
\description{
Prepares the spark environment, downloading the spark binaries
to a common directory if needed.
}

