% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_install}
\alias{spark_connect}
\alias{spark_context}
\alias{spark_disconnect}
\alias{spark_install}
\alias{spark_invoke}
\alias{spark_invoke_static}
\alias{spark_log}
\alias{spark_web}
\title{Provides support to download and install the given Spark version}
\usage{
spark_install(version = "1.6.0")

spark_connect(master = "local", appName = "rspark", version = "1.6.0",
  installInfo = spark_install())

spark_disconnect(sc)

spark_log(con, n = 100)

spark_web(con)

spark_invoke(scon, obj, methodName, ...)

spark_invoke_static(scon, objName, methodName, ...)

spark_context(scon)
}
\description{
Provides support to download and install the given Spark version

Connects to Spark and establishes the Spark Context

Disconnects from Spark and terminates the running application

Prints the last n entries in the Spark log

Opens the Spark web interface

Executes a method on the given object

Executes an static method on the given object

Retrieves the SparkContext reference from a Spark Connection
}

