% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_transactions.R
\docType{methods}
\name{spark-transactions}
\alias{dbBegin,DBISparkConnection-method}
\alias{dbCommit,DBISparkConnection-method}
\alias{dbRollback,DBISparkConnection-method}
\alias{spark-transactions}
\title{Spark Transactions.}
\usage{
\S4method{dbBegin}{DBISparkConnection}(conn)

\S4method{dbCommit}{DBISparkConnection}(conn)

\S4method{dbRollback}{DBISparkConnection}(conn)
}
\description{
\code{dbBegin}, \code{dbCommit} and \code{dbRollback}
are currently not supported and are implementing as no-ops.
}

