% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_context_config.R
\name{spark_advisory_shuffle_partition_size}
\alias{spark_advisory_shuffle_partition_size}
\title{Retrieves or sets advisory size of the shuffle partition}
\usage{
spark_advisory_shuffle_partition_size(sc, size = NULL)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{size}{Advisory size in bytes of the shuffle partition.
Defaults to \code{NULL} to retrieve configuration entries.}
}
\description{
Retrieves or sets advisory size in bytes of the shuffle partition during adaptive optimization
}
\seealso{
Other Spark runtime configuration: 
\code{\link{spark_adaptive_query_execution}()},
\code{\link{spark_auto_broadcast_join_threshold}()},
\code{\link{spark_coalesce_initial_num_partitions}()},
\code{\link{spark_coalesce_min_num_partitions}()},
\code{\link{spark_coalesce_shuffle_partitions}()},
\code{\link{spark_session_config}()}
}
\concept{Spark runtime configuration}
