% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_apply_bundle.R
\name{spark_apply_bundle}
\alias{spark_apply_bundle}
\title{Creates a bundle of dependencies required by \code{spark_apply()}}
\usage{
spark_apply_bundle(packages = TRUE)
}
\arguments{
\item{packages}{List of packages to pack or \code{TRUE} to pack all.}
}
\description{
Creates a bundle of dependencies required by \code{spark_apply()}
}
\keyword{internal}
