% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_apply.R
\name{spark_apply_log}
\alias{spark_apply_log}
\title{Log Writer for Spark Apply}
\usage{
spark_apply_log(..., level = "INFO")
}
\arguments{
\item{...}{Arguments to write to log.}

\item{level}{Severity level for this entry; recommended values: \code{INFO},
\code{ERROR} or \code{WARN}.}
}
\description{
Writes data to log under \code{spark_apply()}.
}
