% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_check_install}
\alias{spark_check_install}
\title{Validates that the given Spark version has been downloaded and installed locally}
\usage{
spark_check_install(version = "1.6.0")
}
\arguments{
\item{version}{Version of Spark to install. Suppported versions: "1.6.0" (default), "2.0.0" (preview)}
}
\description{
Validates that the given Spark version has been downloaded and installed locally
}

