% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_context_config.R
\name{spark_coalesce_shuffle_partitions}
\alias{spark_coalesce_shuffle_partitions}
\title{Retrieves or sets whether coalescing contiguous shuffle partitions is enabled}
\usage{
spark_coalesce_shuffle_partitions(sc, enable = NULL)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{enable}{Whether to enable coalescing of contiguous shuffle partitions.
Defaults to \code{NULL} to retrieve configuration entries.}
}
\description{
Retrieves or sets whether coalescing contiguous shuffle partitions is enabled
}
\seealso{
Other Spark runtime configuration: 
\code{\link{spark_adaptive_query_execution}()},
\code{\link{spark_advisory_shuffle_partition_size}()},
\code{\link{spark_auto_broadcast_join_threshold}()},
\code{\link{spark_coalesce_initial_num_partitions}()},
\code{\link{spark_coalesce_min_num_partitions}()},
\code{\link{spark_session_config}()}
}
\concept{Spark runtime configuration}
