% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_data_frame.R
\name{spark_collect}
\alias{spark_collect}
\title{Read a Spark Dataset into R.}
\usage{
spark_collect(jobj)
}
\arguments{
\item{jobj}{The \code{jobj} underlying a Spark Dataset.}
}
\description{
Read a Spark Dataset into R.
}

