% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{spark_compile}
\alias{spark_compile}
\title{Compiles scala sources and packages into a jar file}
\usage{
spark_compile(name, spark_home)
}
\arguments{
\item{name}{The name of the target jar}

\item{spark_home}{Spark version}
}
\description{
Compiles scala sources and packages into a jar file
}
\keyword{internal}

