% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spark.R
\name{spark_config}
\alias{spark_config}
\title{Defines a configuration file based on the config package and built-in defaults}
\usage{
spark_config(file = "config.yml", use_default = TRUE)
}
\arguments{
\item{file}{Name of the configuration file}

\item{use_default}{TRUE to use the built-in detaults provided in this package}
}
\description{
Defines a configuration file based on the config package and built-in defaults
}

