% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spark.R
\name{spark_config_packages}
\alias{spark_config_packages}
\title{Creates Spark Configuration}
\usage{
spark_config_packages(config, packages, version)
}
\arguments{
\item{config}{The Spark configuration object.}

\item{packages}{A list of named packages or versioned packagese to add.}

\item{version}{The version of Spark being used.}
}
\description{
Creates Spark Configuration
}
\keyword{interenal}
