% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_config.R
\name{spark_config_settings}
\alias{spark_config_settings}
\title{Retrieve Available Settings}
\usage{
spark_config_settings()
}
\description{
Retrieves available sparklyr settings that can be used in configuration files or \code{spark_config()}.
}
