% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{spark_config_value}
\alias{spark_config_value}
\title{A helper function to retrieve values from \code{spark_config()}}
\usage{
spark_config_value(config, name, default = NULL)
}
\arguments{
\item{config}{The configuration list from \code{spark_config()}}

\item{name}{The name of the configuration entry}

\item{default}{The default value to use when entry is not present}
}
\description{
A helper function to retrieve values from \code{spark_config()}
}
\keyword{internal}
