% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_app_name}
\alias{spark_connection_app_name}
\title{Retrieves the application name from a Spark Connection}
\usage{
spark_connection_app_name(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}
}
\description{
Retrieves the application name from a Spark Connection
}

