% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_instances.R
\name{spark_connection_find}
\alias{spark_connection_find}
\title{Find Spark Connection}
\usage{
spark_connection_find(master = NULL, app_name = NULL, method = NULL)
}
\arguments{
\item{master}{The Spark master parameter.}

\item{app_name}{The Spark application name.}

\item{method}{The method used to connect to Spark.}
}
\description{
Finds an active spark connection in the environment given the
connection parameters.
}
