% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_is_local}
\alias{spark_connection_is_local}
\title{TRUE if the Spark Connection is a local install}
\usage{
spark_connection_is_local(sconref)
}
\arguments{
\item{sconref}{Spark connection reference provided by spark_connect}
}
\description{
TRUE if the Spark Connection is a local install
}

