% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_is_open}
\alias{spark_connection_is_open}
\title{Check to see if the connection to Spark is still open}
\usage{
spark_connection_is_open(sc)
}
\arguments{
\item{sc}{Spark connection provided by \code{\link{spark_connect}}}
}
\description{
Check to see if the connection to Spark is still open
}
\seealso{
Other Spark connections: \code{\link{spark_connect}},
  \code{\link{spark_disconnect}}, \code{\link{spark_log}},
  \code{\link{spark_web}}
}
\keyword{internal}

