% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_local_cores}
\alias{spark_connection_local_cores}
\title{Number of cores available in the local install}
\usage{
spark_connection_local_cores(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}
}
\description{
Number of cores available in the local install
}

