% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark_reconnect.R
\name{spark_connection_on_reconnect}
\alias{spark_connection_on_reconnect}
\title{Provides an extension mechanism to allow package builders to support spark_connect(reconnect = TRUE)}
\usage{
spark_connection_on_reconnect(scon, onReconnect)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}

\item{onReconnect}{A function with signature onReconnect(scon, sconInst) that returns an updated sconInst}
}
\description{
Provides an extension mechanism to allow package builders to support spark_connect(reconnect = TRUE)
}

