% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connections_close}
\alias{spark_connection_close_all}
\alias{spark_connections_close}
\title{Closes all existing connections. Returns the total of connections closed.}
\usage{
spark_connection_close_all()
}
\description{
Closes all existing connections. Returns the total of connections closed.
}

