% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_context}
\alias{spark_context}
\title{Retrieves the SparkContext reference from a Spark Connection}
\usage{
spark_context(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}
}
\description{
Retrieves the SparkContext reference from a Spark Connection
}

