% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_context_config.R
\name{spark_context_config}
\alias{spark_context_config}
\title{Runtime configuration interface for Spark.}
\usage{
spark_context_config(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}
}
\description{
Retrieves the runtime configuration interface for Spark.
}
