% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_context_master}
\alias{spark_connection_master}
\alias{spark_context_master}
\title{Retrieves master from a Spark Connection}
\usage{
spark_connection_master(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}
}
\description{
Retrieves master from a Spark Connection
}

