% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dataframe.R
\name{spark_dataframe}
\alias{spark_dataframe}
\title{Get the Spark DataFrame associated with an object}
\usage{
spark_dataframe(x, ...)
}
\arguments{
\item{x}{Object to get DataFrame from}

\item{...}{Reserved for future use}
}
\value{
Reference to DataFrame
}
\description{
S3 method to get the Spark DataFrame associated with objects of
various types.
}

