% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_data_frame.R
\name{spark_dataframe_collect}
\alias{spark_dataframe_collect}
\title{Read a Spark Dataset into R.}
\usage{
spark_dataframe_collect(object)
}
\arguments{
\item{object}{An (object convertable to a) Spark DataFrame.}
}
\description{
Read a Spark Dataset into R.
}

