% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_data_frame.R
\name{spark_dataframe_split}
\alias{spark_dataframe_split}
\title{Split a Spark DataFrame}
\usage{
spark_dataframe_split(object, weights = c(0.5, 0.5),
  seed = sample(.Machine$integer.max, 1))
}
\arguments{
\item{object}{An (object convertable to a) Spark DataFrame.}

\item{weights}{A numeric vector of weights.}

\item{seed}{A numeric seed.}
}
\description{
Split a Spark DataFrame
}

