% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_utils.R
\name{spark_debug_string}
\alias{spark_debug_string}
\title{Description of dependencies for debugging}
\usage{
spark_debug_string(x, print = TRUE)
}
\arguments{
\item{x}{An \R object wrapping, or containing, a Spark DataFrame.}
}
\description{
Print a plan of execution to generate \code{x}. This plan will, among other things, show the
number of partitions in parenthesis at the far left and indicate stages using indendatations.
}
