% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_default_version}
\alias{spark_default_version}
\title{determine the version that will be used by default if version is NULL}
\usage{
spark_default_version()
}
\description{
determine the version that will be used by default if version is NULL
}
\keyword{internal}
