% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dependencies.R
\name{spark_dependency}
\alias{spark_dependency}
\title{Define a Spark dependency}
\usage{
spark_dependency(jars = NULL, packages = NULL)
}
\arguments{
\item{jars}{Character vector of full paths to JAR files}

\item{packages}{Character vector of Spark package names}
}
\value{
An object of class \code{spark_dependency}
}
\description{
Define a Spark dependency, which is a specification of JAR files and/or
Spark packages that are required for an extension.
}

