% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_extensions.R
\name{spark_dependency_fallback}
\alias{spark_dependency_fallback}
\title{Fallback to Spark Dependency}
\usage{
spark_dependency_fallback(spark_version, supported_versions)
}
\arguments{
\item{spark_version}{The Spark version being requested in \code{spark_dependencies}.}

\item{supported_versions}{The Spark versions that are supported by this extension.}
}
\value{
A Spark version to use.
}
\description{
Helper function to assist falling back to previous Spark versions.
}
