% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_disconnect}
\alias{spark_disconnect}
\title{Disconnects from Spark and terminates the running application}
\usage{
spark_disconnect(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}
}
\description{
Disconnects from Spark and terminates the running application
}

