% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_disconnect_all}
\alias{spark_disconnect_all}
\title{Closes all existing connections. Returns the total of connections closed.}
\usage{
spark_disconnect_all()
}
\description{
Closes all existing connections. Returns the total of connections closed.
}

