% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_cluster}
\alias{spark_ec2_cluster}
\title{Create a Spark cluster on EC2}
\usage{
spark_ec2_cluster(access_key_id, secret_access_key, pem_file, version = NULL,
  hadoop_version = NULL, cluster_name = "spark",
  instance_type = "m3.medium", region = NULL)
}
\arguments{
\item{access_key_id}{EC2 access key id. Create a new access key from https://console.aws.amazon.com/iam/home?#security_credential}

\item{secret_access_key}{EC2 secret access key.}

\item{pem_file}{Identity file for ssh connections.}

\item{version}{The Spark version to use.}

\item{hadoop_version}{The Hadoop version to use.}

\item{cluster_name}{Name used to identify cluster.}

\item{instance_type}{Type of EC2 instance. Tested with "m3.medium" and "c3.4xlarge".}

\item{region}{The EC2 region to host this cluster.}
}
\description{
Create a Spark cluster on EC2
}

