% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_deploy}
\alias{spark_ec2_deploy}
\title{Install an EC2 Spark cluster}
\usage{
spark_ec2_deploy(cluster, instance_count = 1, copy_dir = NULL,
  verbose = FALSE)
}
\arguments{
\item{cluster}{A collection of parameters required to use the EC2 cluster, initialized with spark_ec2_cluster.}

\item{instance_count}{The total number of EC2 instances to be provisioned.}

\item{copy_dir}{Copies all the contents (recursevely) of the given path into the driver node durint spark_ec2_deploy}

\item{verbose}{Logs verbose information while executing EC2 commands}
}
\description{
This function will install and launch a new an EC2 cluster and download required client components.
Returns a cluster information list to enable further commands.
}

