% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_destroy}
\alias{spark_ec2_destroy}
\title{Deletes an Spark instance in EC2}
\usage{
spark_ec2_destroy(cluster, verbose = FALSE)
}
\arguments{
\item{cluster}{A collection of parameters required to use the EC2 cluster, initialized with spark_ec2_cluster.}

\item{verbose}{Logs verbose information while executing EC2 commands}
}
\description{
Deletes an Spark instance in EC2
}

