% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_login}
\alias{spark_ec2_login}
\title{Logins into Spark in EC2}
\usage{
spark_ec2_login(cluster, verbose = FALSE)
}
\arguments{
\item{cluster}{A collection of parameters required to use the EC2 cluster, initialized with spark_ec2_cluster.}

\item{verbose}{Logs verbose information while executing EC2 commands}
}
\description{
Logins into Spark in EC2
}

