% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_master}
\alias{spark_ec2_master}
\title{Retrieves master location from EC2}
\usage{
spark_ec2_master(cluster_info, verbose = FALSE)
}
\arguments{
\item{cluster_info}{A collection of parameters required to use the EC2 cluster, initialized with spark_ec2_cluster.}

\item{verbose}{Logs verbose information while executing EC2 commands}
}
\description{
Retrieves master location from EC2
}

