% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_rstudio}
\alias{spark_ec2_rstudio}
\title{Opens RStudio in EC2}
\usage{
spark_ec2_rstudio(cluster)
}
\arguments{
\item{cluster}{A collection of parameters required to use the EC2 cluster, initialized with spark_ec2_cluster.}
}
\description{
Opens RStudio in EC2
}

