% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_spark.R
\name{spark_ec2_start}
\alias{spark_ec2_start}
\title{Starts a previously stopped Spark instance in EC2}
\usage{
spark_ec2_start(cluster, instance_count = 1, verbose = FALSE)
}
\arguments{
\item{cluster}{A collection of parameters required to use the EC2 cluster, initialized with spark_ec2_cluster.}

\item{instance_count}{The total number of EC2 instances to be provisioned.}

\item{verbose}{Logs verbose information while executing EC2 commands}
}
\description{
Starts a previously stopped Spark instance in EC2
}

