% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template.R
\name{spark_extension}
\alias{spark_extension}
\title{Create Spark Extension}
\usage{
spark_extension(path)
}
\arguments{
\item{path}{Location where the extension will be created.}
}
\description{
Creates an R package ready to be used as an Spark extension.
}
