% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{spark_get_java}
\alias{spark_get_java}
\title{Find path to Java}
\usage{
spark_get_java(throws = FALSE)
}
\arguments{
\item{throws}{Throw an error when path not found?}
}
\description{
Finds the path to \code{JAVA_HOME}.
}
\keyword{internal}
