% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_home.R
\name{spark_home_dir}
\alias{spark_home_dir}
\title{Find the SPARK_HOME directory for a version of Spark}
\usage{
spark_home_dir(version = NULL, hadoop_version = NULL)
}
\arguments{
\item{version}{Version of Spark}

\item{hadoop_version}{Version of Hadoop}
}
\value{
Path to SPARK_HOME (or \code{NULL} if the specified version
  was not found).
}
\description{
Find the SPARK_HOME directory for a given version of Spark that
was previously installed using \code{\link{spark_install}}.
}
\keyword{internal}
