% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_home.R
\name{spark_home_set}
\alias{spark_home_set}
\title{Set the SPARK_HOME environment variable}
\usage{
spark_home_set(path = NULL, verbose = getOption("sparklyr.verbose",
  is.null(path)))
}
\arguments{
\item{path}{A string containing the path to the installation location of
Spark. If \code{NULL}, the path to the most latest Spark/Hadoop versions is
used.}

\item{verbose}{Logical. Should the function explain what is it doing?}
}
\value{
The function is mostly invoked for the side-effect of setting the
\code{SPARK_HOME} environment variable. It also returns \code{TRUE} if the
environment was successfully set, and \code{FALSE} otherwise.
}
\description{
Set the \code{SPARK_HOME} environment variable. This slightly speeds up some
operations, including the connection time.
}
\examples{
\dontrun{
# Not run due to side-effects
spark_home_set()
}
}
