% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_install}
\alias{spark_install}
\title{Provides support to download and install the given Spark version}
\usage{
spark_install(version = "1.6.0", reset = FALSE)
}
\arguments{
\item{version}{Version of Spark to install}

\item{reset}{Attempts to reset settings to defaults}
}
\description{
Provides support to download and install the given Spark version
}

