% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_install_available}
\alias{spark_install_available}
\title{Check if the given Spark version is available in this system}
\usage{
spark_install_available(spark_version, hadoop_version)
}
\arguments{
\item{spark_version}{Version of Spark to install. See spark_versions.}

\item{hadoop_version}{Version of Spark to install. See spark_versions_hadoop.}
}
\description{
Check if the given Spark version is available in this system
}

