% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_install_find}
\alias{spark_install_find}
\title{Find a given Spark installation by version.}
\usage{
spark_install_find(
  version = NULL,
  hadoop_version = NULL,
  installed_only = TRUE,
  latest = FALSE,
  hint = FALSE
)
}
\arguments{
\item{version}{Version of Spark to install. See \code{spark_available_versions} for a list of supported versions}

\item{hadoop_version}{Version of Hadoop to install. See \code{spark_available_versions} for a list of supported versions}

\item{installed_only}{Search only the locally installed versions?}

\item{latest}{Check for latest version?}

\item{hint}{On failure should the installation code be provided?}
}
\description{
Find a given Spark installation by version.
}
\keyword{internal}
