% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tools.R
\name{spark_install_sync}
\alias{spark_install_sync}
\title{helper function to sync sparkinstall project to sparklyr}
\usage{
spark_install_sync(project_path)
}
\arguments{
\item{project_path}{The path to the sparkinstall project}
}
\description{
See: https://github.com/rstudio/spark-install
}
\keyword{internal}
