% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_install_tar}
\alias{spark_install_tar}
\title{Provides support to install a version of Spark from a given TAR file}
\usage{
spark_install_tar(path)
}
\arguments{
\item{path}{Path to TAR file conforming to the pattern spark-###-bin-hadoop### where ###
reference spark and hadoop versionsrespectevely.}
}
\description{
Provides support to install a version of Spark from a given TAR file
}

