% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spark_integ_test_skip}
\alias{spark_integ_test_skip}
\title{It lets the package know if it should test a particular functionality or not}
\usage{
spark_integ_test_skip(sc, test_name)
}
\arguments{
\item{sc}{Spark connection}

\item{test_name}{The name of the test}
}
\description{
It lets the package know if it should test a particular functionality or not
}
\details{
It expects a boolean to be returned. If TRUE, the corresponding test will be
skipped. If FALSE the test will be conducted.
}
