% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_invoke_new}
\alias{spark_invoke_new}
\title{Creates a new object of the specified class}
\usage{
spark_invoke_new(sc, class, ...)
}
\arguments{
\item{sc}{Spark connection provided by spark_connect}

\item{class}{Fully-qualified name to static class}

\item{...}{Additional parameters that method requires}
}
\description{
Creates a new object of the specified class
}

