% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_invoke_static_ctor}
\alias{spark_invoke_static_ctor}
\title{Executes an static method on the given object}
\usage{
spark_invoke_static_ctor(sc, class, ...)
}
\arguments{
\item{sc}{Spark connection provided by spark_connect}

\item{class}{Fully-qualified name to static class}

\item{...}{Additional parameters that method requires}
}
\description{
Executes an static method on the given object
}

