% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_jobj.R
\name{spark_jobj}
\alias{spark_jobj}
\title{Get the spark_jobj associated with an object}
\usage{
spark_jobj(x, ...)
}
\arguments{
\item{x}{Object to extract jobj from}

\item{...}{Reserved for future use}
}
\value{
A \code{spark_jobj} object that can be passed to
  \code{\link{invoke}}.
}
\description{
S3 method to get the spark_jobj associated with objects of
various types.
}
\seealso{
\code{\link{invoke}}
}

