% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_invoke.R
\name{spark_last_error}
\alias{spark_last_error}
\title{Surfaces the last error from Spark captured by internal `spark_error` function}
\usage{
spark_last_error()
}
\description{
Surfaces the last error from Spark captured by internal `spark_error` function
}
