% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{spark_load_table}
\alias{spark_load_table}
\title{Reads from a Spark Table into a Spark DataFrame.}
\usage{
spark_load_table(
  sc,
  name,
  path,
  options = list(),
  repartition = 0,
  memory = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{The name to assign to the newly generated table.}

\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3a://"} and \samp{"file://"} protocols.}

\item{options}{A list of strings with additional options. See
\url{https://spark.apache.org/docs/latest/sql-programming-guide.html#configuration}.}

\item{repartition}{The number of partitions used to distribute the
generated table. Use 0 (the default) to avoid partitioning.}

\item{memory}{Boolean; should the data be loaded eagerly into memory? (That
is, should the table be cached?)}

\item{overwrite}{Boolean; overwrite the table with the given name if it
already exists?}
}
\description{
Reads from a Spark Table into a Spark DataFrame.
}
\seealso{
Other Spark serialization routines:
\code{\link[=collect_from_rds]{collect_from_rds()}},
\code{\link[=spark_insert_table]{spark_insert_table()}},
\code{\link[=spark_read]{spark_read()}},
\code{\link[=spark_read_avro]{spark_read_avro()}},
\code{\link[=spark_read_binary]{spark_read_binary()}},
\code{\link[=spark_read_csv]{spark_read_csv()}},
\code{\link[=spark_read_delta]{spark_read_delta()}},
\code{\link[=spark_read_image]{spark_read_image()}},
\code{\link[=spark_read_jdbc]{spark_read_jdbc()}},
\code{\link[=spark_read_json]{spark_read_json()}},
\code{\link[=spark_read_libsvm]{spark_read_libsvm()}},
\code{\link[=spark_read_orc]{spark_read_orc()}},
\code{\link[=spark_read_parquet]{spark_read_parquet()}},
\code{\link[=spark_read_source]{spark_read_source()}},
\code{\link[=spark_read_table]{spark_read_table()}},
\code{\link[=spark_read_text]{spark_read_text()}},
\code{\link[=spark_save_table]{spark_save_table()}},
\code{\link[=spark_write_avro]{spark_write_avro()}},
\code{\link[=spark_write_csv]{spark_write_csv()}},
\code{\link[=spark_write_delta]{spark_write_delta()}},
\code{\link[=spark_write_jdbc]{spark_write_jdbc()}},
\code{\link[=spark_write_json]{spark_write_json()}},
\code{\link[=spark_write_orc]{spark_write_orc()}},
\code{\link[=spark_write_parquet]{spark_write_parquet()}},
\code{\link[=spark_write_source]{spark_write_source()}},
\code{\link[=spark_write_table]{spark_write_table()}},
\code{\link[=spark_write_text]{spark_write_text()}}
}
\concept{Spark serialization routines}
