% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_log}
\alias{spark_log}
\title{Prints the last n entries in the Spark log}
\usage{
spark_log(scon, n = 100)
}
\arguments{
\item{scon}{Spark connection provided by spark_connect}

\item{n}{Max number of log entries to retrieve}
}
\description{
Prints the last n entries in the Spark log
}

