% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_data.R
\name{spark_read_json}
\alias{spark_read_json}
\title{Reads a JSON file and provides a data source compatible with dplyr}
\usage{
spark_read_json(db, name, path, repartition = 0, memory = TRUE,
  overwrite = TRUE)
}
\arguments{
\item{db}{dplyr interface}

\item{name}{Name to reference the data source once it's loaded}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports: "hdfs://" or "s3n://"}

\item{repartition}{Total of partitions used to distribute table or 0 (default) to avoid partitioning}

\item{memory}{Loads data into memory}

\item{overwrite}{Overwrite the table with the given name when it exists}
}
\description{
REads a JSON file and provides a data source compatible with dplyr
}

