% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_spark_data.R
\name{spark_read_parquet}
\alias{spark_read_parquet}
\title{Read a parquet file into a Spark DataFrame}
\usage{
spark_read_parquet(sc, name, path, repartition = 0, memory = TRUE,
  overwrite = TRUE)
}
\arguments{
\item{sc}{The Spark connection}

\item{name}{Name of table}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports: "hdfs://" or "s3n://"}

\item{repartition}{Total of partitions used to distribute table or 0 (default) to avoid partitioning}

\item{memory}{Load data eagerly into memory}

\item{overwrite}{Overwrite the table with the given name if it already exists}
}
\description{
Read a parquet file into a Spark DataFrame
}
\seealso{
Other reading and writing data: \code{\link{spark_read_csv}},
  \code{\link{spark_read_json}},
  \code{\link{spark_write_csv}},
  \code{\link{spark_write_json}},
  \code{\link{spark_write_parquet}}
}

