% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_session}
\alias{spark_session}
\title{Get the Spark Session associated with a connection}
\usage{
spark_session(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}
}
\description{
Get the
\href{http://spark.apache.org/docs/latest/api/scala/#org.apache.spark.sql.SparkSession}{Spark Session}
associated with a \code{spark_connection}.
}
\details{
This object is only available since Spark 2.0.0, and provides
an interface that unifies the Spark SQL context + Hive context
into a single object, and its use is recommended over the older
APIs.
}

