% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{spark_sql}
\alias{spark_sql}
\title{Executes arbitrary SQL statements}
\usage{
spark_sql(con, sql)
}
\arguments{
\item{con}{Connection to dplyr source}

\item{sql}{SQL character string to execute}
}
\description{
Executes arbitrary SQL statements
}

