% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_stat.R
\name{spark_statistical_routines}
\alias{spark_statistical_routines}
\title{Generate random samples from some distribution}
\arguments{
\item{sc}{A Spark connection.}

\item{n}{Sample Size (default: 1000).}

\item{num_partitions}{Number of partitions in the resulting Spark dataframe
(default: default parallelism of the Spark cluster).}

\item{seed}{Random seed (default: a random long integer).}

\item{output_col}{Name of the output column containing sample values (default: "x").}
}
\description{
Generator methods for creating single-column Spark dataframes comprised of
i.i.d. samples from some distribution.
}
