% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_utils.R
\name{spark_table_name}
\alias{spark_table_name}
\title{Generate a Table Name from Expression}
\usage{
spark_table_name(expr)
}
\arguments{
\item{expr}{The expression to attempt to use as name}
}
\description{
Attempts to generate a table name from an expression; otherwise,
assigns an auto-generated generic name with "sparklyr_" prefix.
}
