% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_version.R
\name{spark_version_from_home}
\alias{spark_version_from_home}
\title{Get the Spark Version Associated with a Spark Installation}
\usage{
spark_version_from_home(spark_home, default = NULL)
}
\arguments{
\item{spark_home}{The path to a Spark installation.}

\item{default}{The default version to be inferred, in case
version lookup failed, e.g. no Spark installation was found
at \code{spark_home}.}
}
\description{
Retrieve the version of Spark associated with a Spark installation.
}
