% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions}
\alias{spark_versions}
\title{Retrieves available versions of Spark}
\usage{
spark_versions(latest = TRUE)
}
\arguments{
\item{latest}{TRUE to use the latest download URLs, FALSE to use the package defaults which might be out of date.}
}
\description{
Retrieves available versions of Spark
}

