% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions_download_info}
\alias{spark_versions_download_info}
\title{Retrieves download information for the given Spark and Hadoop versions}
\usage{
spark_versions_download_info(sparkVersion, hadoopVersion)
}
\arguments{
\item{sparkVersion}{The Spark version.}

\item{hadoopVersion}{The Hadoop version.}
}
\description{
Retrieves download information for the given Spark and Hadoop versions
}

