% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions_hadoop}
\alias{spark_versions_hadoop}
\title{Retrieves available versions of Hadoop for Spark}
\usage{
spark_versions_hadoop(spark_version = "1.6.0", supported = TRUE)
}
\arguments{
\item{spark_version}{The Spark version to match for the available hadoop distributions}

\item{supported}{If TRUE (default), returns only the fully supported versions in this package.
Otherwise, retrieves all versions which may be only partially supported.}
}
\value{
Named list of Hadoop versions supported by this version of spark.
}
\description{
Retrieves available versions of Hadoop for Spark
}

