% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions_info}
\alias{spark_versions_info}
\title{Retrieves component information for the given Spark and Hadoop versions}
\usage{
spark_versions_info(spark_version, hadoop_version)
}
\arguments{
\item{spark_version}{The Spark version.}

\item{hadoop_version}{The Hadoop version.}
}
\description{
Retrieves component information for the given Spark and Hadoop versions
}

