% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions_notes_url}
\alias{spark_versions_notes_url}
\title{Opens the release notes for the given version of Spark}
\usage{
spark_versions_notes_url(version)
}
\arguments{
\item{version}{The Spark version. See spark_versions for a list of available options.}
}
\description{
Opens the release notes for the given version of Spark
}

