% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{spark_write}
\alias{spark_write}
\title{Write Spark DataFrame to file using a custom writer}
\usage{
spark_write(x, writer, paths, packages = NULL)
}
\description{
Run a custom R function on Spark worker to write a Spark DataFrame
into file(s)
}
