% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_spark_data.R
\name{spark_write_csv}
\alias{spark_write_csv}
\title{Write a Spark DataFrame to a CSV}
\usage{
spark_write_csv(x, path)
}
\arguments{
\item{x}{A Spark DataFrame or dplyr operation}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports: "hdfs://" or "s3n://"}
}
\description{
Write a Spark DataFrame to a CSV
}
\seealso{
Other reading and writing data: \code{\link{spark_read_csv}},
  \code{\link{spark_read_json}},
  \code{\link{spark_read_parquet}},
  \code{\link{spark_write_json}},
  \code{\link{spark_write_parquet}}
}

