% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{sql_create_table.src_spark}
\alias{sql_create_table.src_spark}
\title{This operation is currently not supported in Spark}
\usage{
sql_create_table.src_spark(...)
}
\arguments{
\item{...}{Additional parameters}
}
\description{
This operation is currently not supported in Spark
}

