% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{sql_drop_table.src_spark}
\alias{sql_drop_table.src_spark}
\title{Removes a Spark table}
\usage{
sql_drop_table.src_spark(con, name)
}
\arguments{
\item{con}{Connection to dplyr source}

\item{name}{Name of the table to remove}
}
\description{
Removes a Spark table
}

