% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{src_context}
\alias{src_context}
\title{Retrieves the Spark connection object from a given dplyr src}
\usage{
src_context(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connection}
}
\description{
Retrieves the Spark connection object from a given dplyr src
}

