% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{src_databases.spark_connection}
\alias{src_databases.spark_connection}
\title{Show database list}
\usage{
src_databases.spark_connection(x, ...)
}
\arguments{
\item{x}{A \code{spark_connection}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Show database list
}
