% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{src_spark}
\alias{src_spark}
\title{Connect to Spark for Dplyr.}
\usage{
src_spark(scon)
}
\arguments{
\item{scon}{Spark connection provided by spark_connection}
}
\description{
Connect to Spark for Dplyr.
}

