% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkster.R
\name{start_local}
\alias{start_local}
\title{Start a local Spark cluster}
\usage{
start_local(version = "1.6.0")
}
\arguments{
\item{version}{The Spark version to use.}
}
\description{
This function will launch the local cluster.
}

