% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_shell.R
\name{start_shell}
\alias{start_shell}
\alias{stop_shell}
\title{Start the Spark R Shell}
\usage{
start_shell(master, spark_home = Sys.getenv("SPARK_HOME"),
  spark_version = NULL, app_name = "sparklyr", config = list(),
  extensions = sparklyr::registered_extensions(), jars = NULL,
  packages = NULL, environment = NULL, shell_args = NULL)

stop_shell(sc)
}
\arguments{
\item{master}{Spark cluster url to connect to. Use \code{"local"} to connect to a local
instance of Spark}

\item{spark_home}{Spark home directory (defaults to SPARK_HOME environment variable)}

\item{spark_version}{Spark version, if not specified, version taken from SPARK_HOME}

\item{app_name}{Application name to be used while running in the Spark cluster}

\item{config}{Named character vector of spark.* options}

\item{extensions}{Extension packages to include dependencies for
(see \code{\link{spark_dependency}}).}

\item{jars}{Paths to Jar files to include}

\item{packages}{Spark packages to include}

\item{environment}{Environment variables to set}

\item{shell_args}{Additional command line arguments for spark_shell}

\item{sc}{\code{spark_connection}}
}
\value{
\code{spark_connection} object
}
\description{
Start the Spark R Shell

Stop the Spark R Shell
}

