% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkster.R
\name{stop_local}
\alias{stop_local}
\title{Stop a local Spark cluster}
\usage{
stop_local(version = "1.6.0")
}
\arguments{
\item{version}{The Spark version to use.}
}
\description{
This function will stop the local cluster.
}

