% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_collect}
\alias{stream_collect}
\title{Collect a Spark Stream}
\usage{
stream_collect(stream)
}
\arguments{
\item{stream}{The spark stream object to collect.}
}
\description{
Collects the given Spark stream by writting snapshot to memory.
}
