% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_find}
\alias{stream_find}
\title{Find Stream}
\usage{
stream_find(sc, id)
}
\arguments{
\item{id}{The stream identifier to find.}
}
\description{
Finds and returns a stream based on the stream's itentifier.
}
