% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_id}
\alias{stream_id}
\title{Spark Stream's Identifier}
\usage{
stream_id(stream)
}
\arguments{
\item{stream}{The spark stream object.}
}
\description{
Retrieves the identifier of the Spark stream.
}
