% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_name}
\alias{stream_name}
\title{Spark Stream's Name}
\usage{
stream_name(stream)
}
\arguments{
\item{stream}{The spark stream object.}
}
\description{
Retrieves the name of the Spark stream if available.
}
