% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_data.R
\name{stream_read_csv}
\alias{stream_read_csv}
\title{Read a CSV stream into a Spark DataFrame}
\usage{
stream_read_csv(sc, name, path)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{The name to assign to the newly generated table.}

\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3a://"} and \samp{"file://"} protocols.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Read a tabular data stream into a Spark DataFrame.
}
\seealso{
Other Spark stream serialization: \code{\link{stream_write_csv}}
}
