% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_data.R
\name{stream_read_delta}
\alias{stream_read_delta}
\title{Read Delta Stream}
\usage{
stream_read_delta(sc, path, name = NULL, options = list(), ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3a://"} and \samp{"file://"} protocols.}

\item{name}{The name to assign to the newly generated stream.}

\item{options}{A list of strings with additional options.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Reads a Delta Lake table as a Spark dataframe stream.
}
\details{
Please note that Delta Lake requires installing the appropriate
package by setting the \code{packages} parameter to \code{"delta"} in \code{spark_connect()}
}
\examples{
\dontrun{

library(sparklyr)
sc <- spark_connect(master = "local", version = "2.4.0", packages = "delta")

sdf_len(sc, 5) \%>\% spark_write_delta(path = "delta-test")

stream <- stream_read_delta(sc, "delta-test") \%>\%
  stream_write_json("json-out")

stream_stop(stream)
}

}
\seealso{
Other Spark stream serialization: 
\code{\link{stream_read_csv}()},
\code{\link{stream_read_json}()},
\code{\link{stream_read_kafka}()},
\code{\link{stream_read_orc}()},
\code{\link{stream_read_parquet}()},
\code{\link{stream_read_socket}()},
\code{\link{stream_read_text}()},
\code{\link{stream_write_console}()},
\code{\link{stream_write_csv}()},
\code{\link{stream_write_delta}()},
\code{\link{stream_write_json}()},
\code{\link{stream_write_kafka}()},
\code{\link{stream_write_memory}()},
\code{\link{stream_write_orc}()},
\code{\link{stream_write_parquet}()},
\code{\link{stream_write_text}()}
}
\concept{Spark stream serialization}
