% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_data.R
\name{stream_read_socket}
\alias{stream_read_socket}
\title{rdname stream_read_csv}
\usage{
stream_read_socket(sc, name = NULL, columns = NULL, options = list(), ...)
}
\description{
rdname stream_read_csv
}
