% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_stop}
\alias{stream_stop}
\title{Stops a Spark Stream}
\usage{
stream_stop(stream)
}
\arguments{
\item{stream}{The spark stream object to be stopped.}
}
\description{
Stops processing data from a Spark stream.
}
