% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_trigger_continuous}
\alias{stream_trigger_continuous}
\title{Spark Stream Continuous Trigger}
\usage{
stream_trigger_continuous(checkpoint = 5000)
}
\arguments{
\item{checkpoint}{The checkpoint interval specified in milliseconds.}
}
\description{
Creates a Spark structured streaming trigger to execute
continuously. This mode is the most performant but not all operations
are supported.
}
\seealso{
\code{\link{stream_trigger_interval}}
}
