% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_trigger_interval}
\alias{stream_trigger_interval}
\title{Spark Stream Interval Trigger}
\usage{
stream_trigger_interval(interval = 1000)
}
\arguments{
\item{interval}{The execution interval specified in milliseconds.}
}
\description{
Creates a Spark structured streaming trigger to execute
over the specified interval.
}
\seealso{
\code{\link{stream_trigger_continuous}}
}
