% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_view.R
\name{stream_view}
\alias{stream_view}
\title{View Stream}
\usage{
stream_view(stream, interval = 1000, ...)
}
\arguments{
\item{stream}{The stream to visualize.}

\item{...}{Additional optional arguments.

#' @examples

library(sparklyr)

stream_read_csv() %>%
  stream_write_csv() %>%
  stream_view(stream) %>%
  stream_stop()}

\item{invalidate}{The invalidation interval in milliseconds.}
}
\description{
Opens a Shiny gadget to visualize the given stream.
}
