% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_operations.R
\name{stream_watermark}
\alias{stream_watermark}
\title{Watermark Stream}
\usage{
stream_watermark(x, column = "timestamp", threshold = "10 minutes")
}
\arguments{
\item{x}{An object coercable to a Spark Streaming DataFrame.}

\item{column}{The name of the column that contains the event time of the row,
if the column is missing, a column with the current time will be added.}

\item{threshold}{The minimum delay to wait to data to arrive late, defaults
to ten minutes.}
}
\description{
Ensures a stream has a watermark defined, which is required for some
operations over streams.
}
