% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_data.R
\name{stream_write_memory}
\alias{stream_write_memory}
\title{Write a Spark DataFrame into Memory}
\usage{
stream_write_memory(x, name = random_string("sparklyr_tmp_"),
  mode = c("append", "complete", "update"),
  trigger = stream_trigger_interval(interval = 5000),
  checkpoint = file.path("checkpoints", name, random_string("")),
  options = list(), ...)
}
\arguments{
\item{x}{A Spark DataFrame or dplyr operation}

\item{name}{The name to use to register this stream.}

\item{mode}{Specifies how data is written to a streaming sink. Valid values are
\code{"append"}, \code{"complete"} or \code{"update"}.}

\item{trigger}{The trigger for the stream query, defaults to micro-batches runnnig
every 5 seconds. See \code{\link{stream_trigger_interval}} and
\code{\link{stream_trigger_continuous}}.}

\item{checkpoint}{The location where the system will write all the checkpoint.
information to guarantee end-to-end fault-tolerance.}

\item{options}{A list of strings with additional options.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Writes a Spark DataFrame into memory.
}
\examples{
\dontrun{

sc <- spark_connect(master = "local")

dir.crate("iris-in")
write.csv(iris, "iris-in/iris.csv", row.names = FALSE)

stream <- stream_read_csv(sc, "iris-in") \%>\% stream_write_memory()

stop_stream(stream)

}

}
\seealso{
Other Spark stream serialization: \code{\link{stream_read_csv}},
  \code{\link{stream_read_memory}},
  \code{\link{stream_write_csv}}
}
