% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_spark.R
\name{tbl_change_db}
\alias{tbl_change_db}
\title{Use specific database}
\usage{
tbl_change_db(sc, name)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{The database name.}
}
\description{
Use specific database
}
