% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{transform_sdf}
\alias{transform_sdf}
\title{transform a subset of column(s) in a Spark Dataframe}
\usage{
transform_sdf(x, cols, fn)
}
\description{
transform a subset of column(s) in a Spark Dataframe
}
