% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_apply.R
\name{worker_spark_apply_unbundle}
\alias{worker_spark_apply_unbundle}
\title{Extracts a bundle of dependencies required by \code{spark_apply()}}
\usage{
worker_spark_apply_unbundle(bundle_path, base_path, bundle_name)
}
\arguments{
\item{bundle_path}{Path to the bundle created using \code{spark_apply_bundle()}}

\item{base_path}{Base path to use while extracting bundles}
}
\description{
Extracts a bundle of dependencies required by \code{spark_apply()}
}
\keyword{internal}
