% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_data_frame.R
\name{as_spark_dataframe}
\alias{as_spark_dataframe}
\title{Convert a tbl to a Spark Dataset.}
\usage{
as_spark_dataframe(x, ...)
}
\arguments{
\item{x}{A \code{spark_tbl}.}
}
\description{
Convert a tbl to a Spark Dataset.
}

