% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{compile_package_jars}
\alias{compile_package_jars}
\title{Compile Scala sources into a Java Archive (jar)}
\usage{
compile_package_jars(package = rprojroot::find_package_root_file(),
  spec = default_compilation_spec())
}
\arguments{
\item{package}{The path to an \R package.}

\item{spec}{The compilation specification, defining how Spark extensions
should be compiled into their resulting Java archives. See
\code{\link{spark_compilation_spec}} for more information.}
}
\description{
Compile the \code{scala} source files contained within an \R package
into a Java Archive (\code{jar}) file that can be loaded and used within
a Spark environment.
}

