% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{copy_to.src_spark}
\alias{copy_to.src_spark}
\title{Copies the source data frame into a Spark table}
\usage{
\method{copy_to}{src_spark}(con, df, name, cache = TRUE, repartition = 16)
}
\arguments{
\item{con}{Connection to dplyr source}

\item{df}{Data frame to copy from}

\item{name}{Name of the destination table}

\item{cache}{Cache table for improved performance}

\item{repartition}{Number of partitions to repartition table}
}
\description{
Copies the source data frame into a Spark table
}

