% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{copy_to.src_spark}
\alias{copy_to.src_spark}
\title{Copies the source data frame into a Spark table}
\usage{
\method{copy_to}{src_spark}(db, df, name)
}
\description{
Copies the source data frame into a Spark table
}

