% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_connection.R
\docType{class}
\name{DBISparkConnection-class}
\alias{DBISparkConnection-class}
\alias{dbConnect,DBISparkDriver-method}
\alias{dbDataType,DBISparkConnection,ANY-method}
\alias{dbDataType,DBISparkDriver,ANY-method}
\alias{dbDisconnect,DBISparkConnection-method}
\alias{dbGetInfo,DBISparkConnection-method}
\alias{dbQuoteIdentifier,DBISparkConnection,character-method}
\alias{show,DBISparkConnection-method}
\title{DBISparkConnection and methods.}
\usage{
\S4method{dbGetInfo}{DBISparkConnection}(dbObj, ...)

\S4method{show}{DBISparkConnection}(object)

\S4method{dbConnect}{DBISparkDriver}(drv, master = NULL, ...)

\S4method{dbDisconnect}{DBISparkConnection}(conn, ...)

\S4method{dbDataType}{DBISparkDriver,ANY}(dbObj, obj)

\S4method{dbDataType}{DBISparkConnection,ANY}(dbObj, obj)

\S4method{dbQuoteIdentifier}{DBISparkConnection,character}(conn, x, ...)
}
\arguments{
\item{dbObj}{Spark driver or connection.}

\item{drv}{\code{splyr::DBISpark()}}

\item{master}{Master location.}

\item{obj}{Object to convert}
}
\description{
DBISparkConnection and methods.

Connect to Spark

Determine database type for R vector.
}
\examples{
library(DBI)
con <- dbConnect(splyr::DBISpark())
dbDisconnect(con)
}
\keyword{internal}

