% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_spark_connection.R
\name{dbSetProperty}
\alias{dbSetProperty}
\title{Sets a property for the connection}
\usage{
dbSetProperty(conn, property, value)
}
\description{
Sets a property for the connection
}
\keyword{internal}

