% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{download_scalac}
\alias{download_scalac}
\title{Downloads default Scala Compilers}
\usage{
download_scalac(dest_path = NULL)
}
\arguments{
\item{dest_path}{The destination path where scalac will be
downloaded to.}
}
\description{
\code{compile_package_jars} requires several versions of the
scala compiler to work, this is to match Spark scala versions.
To help setup your environment, this function will download the
required compilers under the default search path.
}
\details{
See \code{find_scalac} for a list of paths searched and used by
this function to install the required compilers.
}
