% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_table.R
\name{dplyr-spark-table}
\alias{collect.tbl_spark}
\alias{dplyr-spark-table}
\alias{sample_frac.tbl_spark}
\alias{sample_n.tbl_spark}
\alias{sql_build.tbl_spark}
\title{Dplyr table definitions for Spark}
\usage{
\method{collect}{tbl_spark}(x, ..., n = 1e+05, warn_incomplete = TRUE)

sql_build.tbl_spark(op, con, ...)

\method{sample_n}{tbl_spark}(.data, size, replace = FALSE, weight = NULL,
  .env = parent.frame(), ..., .dots)

\method{sample_frac}{tbl_spark}(.data, size = 1, replace = FALSE,
  weight = NULL, .env = parent.frame(), ..., .dots)
}
\description{
Dplyr table definitions for Spark
}

