% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_transformation.R
\name{ft_sql_transformer}
\alias{ft_sql_transformer}
\title{Feature Transformation -- SQLTransformer}
\usage{
ft_sql_transformer(x, input_col = NULL, output_col = NULL, sql)
}
\arguments{
\item{x}{An object (usually a \code{spark_tbl}) coercable to a Spark DataFrame.}

\item{input_col}{The name of the input column(s).}

\item{output_col}{The name of the output column.}

\item{sql}{A SQL statement.}
}
\description{
Transform a data set using SQL. Use the \code{__THIS__}
placeholder as a proxy for the active table.
}
\seealso{
See \url{http://spark.apache.org/docs/latest/ml-features.html} for
  more information on the set of transformations available for DataFrame
  columns in Spark.

Other feature transformation routines: \code{\link{ft_binarizer}},
  \code{\link{ft_bucketizer}},
  \code{\link{ft_discrete_cosine_transform}},
  \code{\link{ft_elementwise_product}},
  \code{\link{ft_index_to_string}},
  \code{\link{ft_one_hot_encoder}},
  \code{\link{ft_quantile_discretizer}},
  \code{\link{ft_string_indexer}},
  \code{\link{ft_vector_assembler}},
  \code{\link{sdf_mutate}}
}

