% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_hof.R
\name{hof_aggregate}
\alias{hof_aggregate}
\title{dplyr wrapper for \code{aggregate(array<T>, A, function<A, T, A>[, function<A, R>]): R}}
\usage{
hof_aggregate(x, dest_col, expr, start, merge, ...)
}
\arguments{
\item{x}{The Spark data frame to run aggregation on}

\item{dest_col}{Column to store the aggregated result}

\item{expr}{The array being aggregated (could be any SQL expression evaluating to an array)}

\item{start}{The starting value of the aggregation}

\item{merge}{The aggregation function}

\item{...}{Additional params to dplyr::mutate}

\item{finish}{Optional param specifying a transformation to apply on the final value of the aggregation}
}
\description{
dplyr wrapper for \code{aggregate(array<T>, A, function<A, T, A>[, function<A, R>]): R}
}
