% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_hof.R
\name{hof_exists}
\alias{hof_exists}
\title{Determine Whether Some Element Exists in an Array Column}
\usage{
hof_exists(x, pred, expr = NULL, dest_col = NULL, ...)
}
\arguments{
\item{x}{The Spark data frame to search}

\item{pred}{A boolean predicate}

\item{expr}{The array being searched (could be any SQL expression evaluating to an array)}

\item{dest_col}{Column to store the search result}

\item{...}{Additional params to dplyr::mutate}
}
\description{
Determines whether an element satisfying the given predicate exists in each array from
an array column
(this is essentially a dplyr wrapper for the
\code{exists(array<T>, function<T, Boolean>): Boolean} built-in Spark SQL function)
}
