% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_livy.R
\name{livy_install}
\alias{livy_install}
\title{Install Livy}
\usage{
livy_install(spark_home = Sys.getenv("SPARK_HOME"), livy_version = "0.2.0",
  hadoop_conf_dir = NULL)
}
\arguments{
\item{spark_home}{The path to a Spark installation. The downloaded and
installed version of \samp{livy} will then be associated with this Spark
installation.}

\item{livy_version}{The version of \samp{livy} to be installed. Currently,
only the beta version, 0.2.0, is available.}

\item{hadoop_conf_dir}{Optional; the path to a Hadoop configuration directory.}
}
\description{
Automatically download and install \href{http://livy.io/}{\samp{livy}}.
\samp{livy} provides a REST API to Spark.
}

