% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livy_service.R
\name{livy_service_start}
\alias{livy_service_start}
\alias{livy_service_stop}
\title{Start Livy}
\usage{
livy_service_start(version = NULL, spark_version = NULL, stdout = "",
  stderr = "", ...)

livy_service_stop()
}
\arguments{
\item{version}{The version of \samp{livy} to use.}

\item{spark_version}{The version of \samp{spark} to connect to.}

\item{stdout, stderr}{where output to 'stdout' or 'stderr' should
be sent. Same options as \code{system2}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Starts the livy service.

Stops the running instances of the livy service.
}
