% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_transformer_and_estimator.R
\name{ml-constructors}
\alias{ml-constructors}
\alias{new_ml_transformer}
\alias{new_ml_prediction_model}
\alias{new_ml_classification_model}
\alias{new_ml_probabilistic_classification_model}
\alias{new_ml_clustering_model}
\alias{new_ml_estimator}
\alias{new_ml_predictor}
\alias{new_ml_classifier}
\alias{new_ml_probabilistic_classifier}
\title{Constructors for Pipeline Stages}
\usage{
new_ml_transformer(jobj, ..., class = character())

new_ml_prediction_model(jobj, ..., class = character())

new_ml_classification_model(jobj, ..., class = character())

new_ml_probabilistic_classification_model(jobj, ..., class = character())

new_ml_clustering_model(jobj, ..., class = character())

new_ml_estimator(jobj, ..., class = character())

new_ml_predictor(jobj, ..., class = character())

new_ml_classifier(jobj, ..., class = character())

new_ml_probabilistic_classifier(jobj, ..., class = character())
}
\arguments{
\item{jobj}{Pointer to the pipeline stage object.}

\item{...}{(Optional) additional attributes of the object.}

\item{class}{Name of class.}
}
\description{
Functions for developers writing extensions for Spark ML.
}
\keyword{internal}
