% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_kmeans.R
\name{ml_kmeans}
\alias{ml_kmeans}
\title{Spark ML -- K-Means Clustering}
\usage{
ml_kmeans(x, centers, iter.max = 10, features = dplyr::tbl_vars(x))
}
\arguments{
\item{x}{A \code{spark_tbl}.}

\item{centers}{The number of cluster centers to compute.}

\item{iter.max}{Maximum number of iterations allowed.}

\item{features}{Which columns to use in the k-means fit. Defaults to
all columns within \code{x}.}
}
\description{
Perform k-means clustering on a \code{spark_tbl}.
}
\seealso{
For information on how Spark k-means clustering is implemented, please see
  \url{http://spark.apache.org/docs/latest/mllib-clustering.html#k-means}.

Other Spark ML routines: \code{\link{ml_linear_regression}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_multilayer_perceptron}},
  \code{\link{ml_pca}}, \code{\link{ml_random_forest}}
}

