% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_clustering_kmeans.R
\name{ml_kmeans}
\alias{ml_kmeans}
\title{Spark ML -- K-Means Clustering}
\usage{
ml_kmeans(x, k = 2L, max_iter = 20L, tol = 1e-04, init_steps = 2L,
  init_mode = "k-means||", seed = NULL, features_col = "features",
  prediction_col = "prediction", uid = random_string("kmeans_"), ...)
}
\description{
K-means clustering with support for k-means|| initialization proposed by Bahmani et al.
}
