% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_kmeans.R
\name{ml_kmeans}
\alias{ml_kmeans}
\title{Spark ML -- K-Means Clustering}
\usage{
ml_kmeans(x, centers, iter.max = 100, features = tbl_vars(x),
  compute.cost = TRUE, tolerance = 1e-04, ml.options = ml_options(), ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{centers}{The number of cluster centers to compute.}

\item{iter.max}{The maximum number of iterations to use.}

\item{features}{The name of features (terms) to use for the model fit.}

\item{compute.cost}{Whether to compute cost for \code{k-means} model using Spark's \href{https://github.com/apache/spark/blob/master/mllib/src/main/scala/org/apache/spark/mllib/clustering/KMeansModel.scala#L84}{computeCost}.}

\item{ml.options}{Optional arguments, used to affect the model generated. See
\code{\link{ml_options}} for more details.}

\item{...}{Optional arguments; currently unused.}

\item{tol}{Param for the convergence tolerance for iterative algorithms.}
}
\value{
\link{ml_model} object of class \code{kmeans} with overloaded \code{print}, \code{fitted} and \code{predict} functions.
}
\description{
Perform k-means clustering on a Spark DataFrame.
}
\references{
Bahmani et al., Scalable K-Means++, VLDB 2012
}
\seealso{
For information on how Spark k-means clustering is implemented, please see
  \url{http://spark.apache.org/docs/latest/mllib-clustering.html#k-means}.
}
