% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_lda.R
\name{ml_lda}
\alias{ml_lda}
\title{Spark ML -- Latent Dirichlet Allocation}
\usage{
ml_lda(x, features = dplyr::tbl_vars(x), ...)
}
\arguments{
\item{x}{An object convertable to a Spark DataFrame (typically, a \code{tbl_spark}).}

\item{features}{The columns to use in the principal components
analysis. Defaults to all columns in \code{x}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Fit a Latent Dirichlet Allocation (LDA) model to a Spark DataFrame.
}
\seealso{
Other Spark ML routines: \code{\link{ml_decision_tree}},
  \code{\link{ml_gradient_boosted_trees}},
  \code{\link{ml_kmeans}},
  \code{\link{ml_linear_regression}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_multilayer_perceptron}},
  \code{\link{ml_naive_bayes}},
  \code{\link{ml_one_vs_rest}}, \code{\link{ml_pca}},
  \code{\link{ml_random_forest}},
  \code{\link{ml_survival_regression}}
}

