% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_one_vs_rest.R
\name{ml_one_vs_rest}
\alias{ml_one_vs_rest}
\title{Spark ML -- One vs Rest}
\usage{
ml_one_vs_rest(x, classifier, response, features)
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame (typically, a
\code{tbl_spark}).}

\item{classifier}{The classifier model. Can be obtained using the \code{only_model} parameter.}

\item{response}{The name of the response vector (as a length-one character
vector), or a formula, giving a symbolic description of the model to be
fitted. When \code{response} is a formula, it is used in preference to other
parameters to set the \code{response}, \code{features}, and \code{intercept}
parameters (if available).}

\item{features}{The name of features (terms) to use for the model fit.}
}
\description{
Perform regression or classification using one vs rest.
}
\seealso{
Other Spark ML routines: \code{\link{ml_decision_tree}},
  \code{\link{ml_generalized_linear_regression}},
  \code{\link{ml_gradient_boosted_trees}},
  \code{\link{ml_kmeans}}, \code{\link{ml_lda}},
  \code{\link{ml_linear_regression}},
  \code{\link{ml_logistic_regression}},
  \code{\link{ml_multilayer_perceptron}},
  \code{\link{ml_naive_bayes}}, \code{\link{ml_pca}},
  \code{\link{ml_random_forest}},
  \code{\link{ml_survival_regression}}
}

