% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_options.R
\name{ml_options}
\alias{ml_options}
\title{Provide Options for Spark.ML Routines}
\usage{
ml_options(feature.col, ...)
}
\arguments{
\item{feature.col}{The name to assign to the feature column.}

\item{Optional}{arguments, reserved for future expansion.}
}
\description{
Provide Options for Spark.ML Routines
}

