% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model_print_methods.R
\name{ml_tree_feature_importance}
\alias{ml_tree_feature_importance}
\title{Spark ML - Feature Importance for Tree Models}
\usage{
ml_tree_feature_importance(sc, model)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{model}{An \code{ml_model} encapsulating the output from a decision tree.}
}
\value{
A sorted data frame with feature labels and their relative importance.
}
\description{
Spark ML - Feature Importance for Tree Models
}

