% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_print_utils.R
\name{ml_tree_feature_importance}
\alias{ml_tree_feature_importance}
\title{Spark ML - Feature Importance for Tree Models}
\usage{
ml_tree_feature_importance(model, ...)
}
\arguments{
\item{model}{A decision tree-based \code{ml_model}}

\item{...}{Optional arguments; currently unused.}
}
\value{
A sorted data frame with feature labels and their relative importance.
}
\description{
Spark ML - Feature Importance for Tree Models
}
