% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers_ml_tree_models.R
\name{ml_tree_tidiers}
\alias{ml_tree_tidiers}
\alias{tidy.ml_model_decision_tree_classification}
\alias{tidy.ml_model_decision_tree_regression}
\alias{augment.ml_model_decision_tree_classification}
\alias{augment._ml_model_decision_tree_classification}
\alias{augment.ml_model_decision_tree_regression}
\alias{augment._ml_model_decision_tree_regression}
\alias{glance.ml_model_decision_tree_classification}
\alias{glance.ml_model_decision_tree_regression}
\alias{tidy.ml_model_random_forest_classification}
\alias{tidy.ml_model_random_forest_regression}
\alias{augment.ml_model_random_forest_classification}
\alias{augment._ml_model_random_forest_classification}
\alias{augment.ml_model_random_forest_regression}
\alias{augment._ml_model_random_forest_regression}
\alias{glance.ml_model_random_forest_classification}
\alias{glance.ml_model_random_forest_regression}
\alias{tidy.ml_model_gbt_classification}
\alias{tidy.ml_model_gbt_regression}
\alias{augment.ml_model_gbt_classification}
\alias{augment._ml_model_gbt_classification}
\alias{augment.ml_model_gbt_regression}
\alias{augment._ml_model_gbt_regression}
\alias{glance.ml_model_gbt_classification}
\alias{glance.ml_model_gbt_regression}
\title{Tidying methods for Spark ML tree models}
\usage{
\method{tidy}{ml_model_decision_tree_classification}(x, ...)

\method{tidy}{ml_model_decision_tree_regression}(x, ...)

\method{augment}{ml_model_decision_tree_classification}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_decision_tree_classification`}(x, new_data = NULL, ...)

\method{augment}{ml_model_decision_tree_regression}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_decision_tree_regression`}(x, new_data = NULL, ...)

\method{glance}{ml_model_decision_tree_classification}(x, ...)

\method{glance}{ml_model_decision_tree_regression}(x, ...)

\method{tidy}{ml_model_random_forest_classification}(x, ...)

\method{tidy}{ml_model_random_forest_regression}(x, ...)

\method{augment}{ml_model_random_forest_classification}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_random_forest_classification`}(x, new_data = NULL, ...)

\method{augment}{ml_model_random_forest_regression}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_random_forest_regression`}(x, new_data = NULL, ...)

\method{glance}{ml_model_random_forest_classification}(x, ...)

\method{glance}{ml_model_random_forest_regression}(x, ...)

\method{tidy}{ml_model_gbt_classification}(x, ...)

\method{tidy}{ml_model_gbt_regression}(x, ...)

\method{augment}{ml_model_gbt_classification}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_gbt_classification`}(x, new_data = NULL, ...)

\method{augment}{ml_model_gbt_regression}(x, newdata = NULL, ...)

\method{augment}{`_ml_model_gbt_regression`}(x, new_data = NULL, ...)

\method{glance}{ml_model_gbt_classification}(x, ...)

\method{glance}{ml_model_gbt_regression}(x, ...)
}
\arguments{
\item{x}{a Spark ML model.}

\item{...}{extra arguments (not used.)}

\item{newdata}{a tbl_spark of new data to use for prediction.}

\item{new_data}{a tbl_spark of new data to use for prediction.}
}
\description{
These methods summarize the results of Spark ML models into tidy forms.
}
