% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_spark.R
\name{options-spark}
\alias{options-spark}
\title{The rspark.dplyr.optimize_shuffle_cores option optimizes the number of partitions
to use when shuffling data based on the number of local cores.
This option is only applicable to local installations. The default TRUE provides
automatic detection of cores. To avoid using this setting, set this value to FALSE.}
\description{
The rspark.dplyr.optimize_shuffle_cores option optimizes the number of partitions
to use when shuffling data based on the number of local cores.
This option is only applicable to local installations. The default TRUE provides
automatic detection of cores. To avoid using this setting, set this value to FALSE.

The rspark.packages.default option, overrides the default Spark packages

The rspark.install.dir option, defines the default Spark installation directory

The spark.connection.allow_local_reconnect option, allows reconnect to
be enabled in local installations to help troubleshoot remote environments.

The spark.connection.codegen option, sets the spark.sql.codegen.wholeStage option.
}
\examples{
getOption("rspark.dplyr.optimize_shuffle_cores", TRUE)
getOption("rspark.packages.default", NULL)
getOption("rspark.install.dir", TRUE)
getOption("spark.connection.allow_local_reconnect", FALSE)
getOption("spark.connection.codegen", FALSE)
}

