% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_spark_jobj.R
\name{print.jobj}
\alias{print.jobj}
\title{Print a JVM object reference.}
\usage{
\method{print}{jobj}(x, ...)
}
\arguments{
\item{x}{The JVM object reference}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function prints the type and id for an object stored
in the Spark R API JVM backend.
}

