% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{augment}
\alias{glance}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{full_join}
\alias{nest}
\alias{unnest}
\alias{separate}
\alias{unite}
\alias{pivot_wider}
\alias{pivot_longer}
\alias{fill}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr:mutate-joins]{right_join}}}

  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{tidyr}{\code{\link[tidyr]{fill}}, \code{\link[tidyr]{nest}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}, \code{\link[tidyr:nest]{unnest}}}
}}

