% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_extensions.R
\name{register_extension}
\alias{register_extension}
\alias{registered_extensions}
\title{Register a Package that Implements a Spark Extension}
\usage{
register_extension(package)

registered_extensions()
}
\arguments{
\item{package}{The package(s) to register.}
}
\description{
Registering an extension package will result in the package being
automatically scanned for spark dependencies when a connection to Spark is
created.
}
\note{
Packages should typically register their extensions in their
  \code{.onLoad} hook -- this ensures that their extensions are registered
  when their namespaces are loaded.
}
