% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_spark.R
\name{registerDoSpark}
\alias{registerDoSpark}
\title{Register a Prallel Backend}
\usage{
registerDoSpark(spark_conn, min_executors = 0L, ...)
}
\arguments{
\item{spark_conn}{spark connection to use}

\item{min_executors}{Minimum number of Spark executors that is expected to be
available}

\item{...}{additional options for sparklyr parallel backend
(currently only the only valid option is nocompile = {T, F})}
}
\value{
None
}
\description{
Registers a parallel backend using the \code{foreach} package.
}
\examples{
\dontrun{

sc <- spark_connect(master = "local")
registerDoSpark(sc, nocompile = FALSE)
}

}
