% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_sequence.R
\name{sdf_along}
\alias{sdf_along}
\title{Create DataFrame for along Object}
\usage{
sdf_along(sc, along, repartition = NULL)
}
\arguments{
\item{sc}{The associated Spark connection.}

\item{along}{Takes the length from the length of this argument.}

\item{repartition}{The number of partitions to use when distributing the
data across the Spark cluster.}
}
\description{
Creates a DataFrame along the given object.
}
