% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_wrapper.R
\name{sdf_collect}
\alias{sdf_collect}
\title{Collect a Spark DataFrame into R.}
\usage{
sdf_collect(object, ...)
}
\arguments{
\item{...}{Additional options.}

\item{Spark}{Spark dataframe to collect}
}
\description{
Collects a Spark dataframe into R.
}
