% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_distinct.R
\name{sdf_distinct}
\alias{sdf_distinct}
\title{Invoke distinct on a Spark DataFrame}
\usage{
sdf_distinct(x, ..., name)
}
\arguments{
\item{x}{A Spark DataFrame.}

\item{...}{Optional variables to use when determining uniqueness.
If there are multiple rows for a given combination of inputs,
only the first row will be preserved. If omitted, will use all
variables.}

\item{name}{A name to assign this table. Passed to [sdf_register()].}
}
\description{
Invoke distinct on a Spark DataFrame
}
\seealso{
Other Spark data frames: 
\code{\link{sdf_copy_to}()},
\code{\link{sdf_random_split}()},
\code{\link{sdf_register}()},
\code{\link{sdf_sample}()},
\code{\link{sdf_sort}()},
\code{\link{sdf_weighted_sample}()}
}
\concept{Spark data frames}
