% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_last_index}
\alias{sdf_last_index}
\title{Returns the last index of a Spark DataFrame}
\usage{
sdf_last_index(x, id = "id")
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{id}{The name of the index column.}
}
\description{
Returns the last index of a Spark DataFrame. The Spark
\code{mapPartitionsWithIndex} function is used to iterate
through the last nonempty partition of the RDD to find the last record.
}
