% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_predict}
\alias{sdf_predict}
\title{Model Predictions with Spark DataFrames}
\usage{
sdf_predict(object, newdata, ...)
}
\arguments{
\item{object, newdata}{An object coercable to a Spark DataFrame.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Given a \code{ml_model} fit alongside a new data set, produce a new Spark
DataFrame with predicted values encoded in the \code{"prediction"} column.
}
\seealso{
Other Spark data frames: \code{\link{sdf_copy_to}},
  \code{\link{sdf_partition}}, \code{\link{sdf_register}},
  \code{\link{sdf_sample}}, \code{\link{sdf_sort}}
}
