% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_register}
\alias{sdf_register}
\title{Register a Spark DataFrame}
\usage{
sdf_register(x, name = NULL)
}
\arguments{
\item{x}{A Spark DataFrame.}

\item{name}{A name to assign this table.}
}
\description{
Registers a Spark DataFrame (giving it a table name for the
Spark SQL context), and returns a \code{tbl_spark}.
}
\seealso{
Other Spark data frames: 
\code{\link{sdf_copy_to}()},
\code{\link{sdf_distinct}()},
\code{\link{sdf_random_split}()},
\code{\link{sdf_sample}()},
\code{\link{sdf_sort}()},
\code{\link{sdf_weighted_sample}()}
}
\concept{Spark data frames}
