% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_sample}
\alias{sdf_sample}
\title{Randomly Sample Rows from a Spark DataFrame}
\usage{
sdf_sample(x, fraction = 1, replacement = TRUE, seed = NULL)
}
\arguments{
\item{x}{An object coercable to a Spark DataFrame.}

\item{fraction}{The fraction to sample.}

\item{replacement}{Boolean; sample with replacement?}

\item{seed}{An (optional) integer seed.}
}
\description{
Draw a random sample of rows (with or without replacement)
from a Spark DataFrame.
}
\seealso{
Other Spark data frames: 
\code{\link{sdf_copy_to}()},
\code{\link{sdf_distinct}()},
\code{\link{sdf_random_split}()},
\code{\link{sdf_register}()},
\code{\link{sdf_sort}()},
\code{\link{sdf_weighted_sample}()}
}
\concept{Spark data frames}
