% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_sequence.R
\name{sdf_seq}
\alias{sdf_seq}
\title{Create DataFrame for Range}
\usage{
sdf_seq(sc, from = 1L, to = 1L, by = 1L, repartition = NULL)
}
\arguments{
\item{sc}{The associated Spark connection.}

\item{from, to}{The start and end to use as a range}

\item{by}{The increment of the sequence.}

\item{repartition}{The number of partitions to use when distributing the
data across the Spark cluster.}
}
\description{
Creates a DataFrame for the given range
}
