% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_check_install}
\alias{spark_check_install}
\title{Validates that the given Spark version has been downloaded and installed locally}
\usage{
spark_check_install(spark_version = "1.6.0", hadoop_version = "2.6")
}
\arguments{
\item{spark_version}{Version of Spark to install. See spark_versions.}

\item{hadoop_version}{Version of Spark to install. See spark_versions_hadoop.}
}
\description{
Validates that the given Spark version has been downloaded and installed locally
}

