% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{spark_compile}
\alias{spark_compile}
\title{Compiles scala sources and packages into a jar file}
\usage{
spark_compile(name, spark_version = "1.6.1", hadoop_version = "2.6")
}
\arguments{
\item{name}{The name of the target jar}

\item{spark_version}{Spark version}

\item{hadoop_version}{Hadoop version}
}
\description{
Compiles scala sources and packages into a jar file
}
\keyword{internal}

