% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_compile.R
\name{spark_compile}
\alias{spark_compile}
\title{Compile Scala sources into a Java Archive (jar)}
\usage{
spark_compile(name, spark_home, scalac = NULL, jar = NULL)
}
\arguments{
\item{name}{The name to assign to the target \code{jar}.}

\item{spark_home}{The path to the Spark sources to be used
alongside compilation.}

\item{scalac}{The path to the \code{scalac} program to be used, for
compilation of \code{scala} files.}

\item{jar}{The path to the \code{jar} program to be used, for
generating of the resulting \code{jar}.}
}
\description{
Given a set of \code{scala} source files, compile them
into a Java Archive (\code{jar}).
}
\keyword{internal}

