% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spark.R
\name{spark_config}
\alias{spark_config}
\title{Read Spark Configuration}
\usage{
spark_config(file = "config.yml", use_default = TRUE)
}
\arguments{
\item{file}{Name of the configuration file}

\item{use_default}{TRUE to use the built-in defaults provided in this package}
}
\value{
Named list with configuration data
}
\description{
Read Spark Configuration
}
\details{
Read Spark configuration using the \pkg{\link[config]{config}} package.
}
