% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connect}
\alias{spark_connect}
\title{Connect to Spark}
\usage{
spark_connect(master = "local", app_name = "sparklyr", version = NULL,
  hadoop_version = NULL, extensions = NULL, config = spark_config())
}
\arguments{
\item{master}{Master definition to Spark cluster}

\item{app_name}{Application name to be used while running in the Spark cluster}

\item{version}{Version of the Spark (only applicable for local mode)}

\item{hadoop_version}{Version of Hadoop (only applicable for local mode)}

\item{extensions}{Extension packages to enable for this connection.}

\item{config}{Configuration for connection (see \code{\link{spark_config} for details}).}
}
\value{
Connection to Spark local instance or remote cluster
}
\description{
Connect to Spark
}
\seealso{
Other Spark connections: \code{\link{spark_connection_is_open}},
  \code{\link{spark_disconnect}}, \code{\link{spark_log}},
  \code{\link{spark_web}}
}

