% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connect}
\alias{spark_connect}
\title{Connects to Spark and establishes the Spark Context}
\usage{
spark_connect(master = "local", appName = "rspark", version = "1.6.0",
  installInfo = spark_install())
}
\arguments{
\item{master}{Master definition to Spark cluster}

\item{appName}{Application name to be used while running in the Spark cluster}

\item{version}{Version of the Spark cluster}

\item{installInfo}{Description of installation paths}
}
\description{
Connects to Spark and establishes the Spark Context
}

