% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_connection}
\alias{spark_connection}
\title{Get the spark_connection associated with an object}
\usage{
spark_connection(x, ...)
}
\arguments{
\item{x}{Object to extract connection from}

\item{...}{Reserved for future use}
}
\value{
A \code{spark_connection} object that can be passed to
  \code{\link{invoke_new}} and \code{\link{invoke_static}}.
}
\description{
S3 method to get the spark_connection associated with objects of
various types.
}

