% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_connection}
\alias{spark_connection}
\title{Retrieve the Spark Connection Associated with an R Object}
\usage{
spark_connection(x, ...)
}
\arguments{
\item{x}{An \R object from which a \code{spark_connection} can be obtained.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Retrieve the \code{spark_connection} associated with an \R object.
}
