% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_cores}
\alias{spark_connection_cores}
\title{Number of cores available in the cluster}
\usage{
spark_connection_cores(sconref)
}
\arguments{
\item{sconref}{Spark connection reference provided by spark_connect}
}
\description{
Number of cores available in the cluster
}

