% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_is_open}
\alias{spark_connection_is_open}
\title{Checks to see if the connection into Spark is still open}
\usage{
spark_connection_is_open(sconref)
}
\arguments{
\item{sconref}{Spark connection reference provided by spark_connect}
}
\description{
Checks to see if the connection into Spark is still open
}

