% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_connection_is_open}
\alias{spark_connection_is_open}
\title{Check if a Spark connection is open}
\usage{
spark_connection_is_open(sc)
}
\arguments{
\item{sc}{\code{spark_connection}}
}
\description{
Check if a Spark connection is open
}
\keyword{internal}

