% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_context}
\alias{spark_context}
\title{Get the SparkContext associated with a connection}
\usage{
spark_context(sc)
}
\arguments{
\item{sc}{Connection to get SparkContext from}
}
\value{
Reference to SparkContext
}
\description{
Get the SparkContext \code{spark_jobj} associated with a
\code{spark_connection}
}

