% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_dataframe.R
\name{spark_dataframe}
\alias{spark_dataframe}
\title{Retrieve a Spark DataFrame}
\usage{
spark_dataframe(x, ...)
}
\arguments{
\item{x}{An \R object wrapping, or containing, a Spark DataFrame.}

\item{...}{Optional arguments; currently unused.}
}
\value{
A \code{\link{spark_jobj}} representing a Java object reference
  to a Spark DataFrame.
}
\description{
This S3 generic is used to access a Spark DataFrame object (as a Java
object reference) from an \R object.
}
