% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_extensions.R
\name{spark_dependency}
\alias{spark_dependency}
\title{Define a Spark dependency}
\usage{
spark_dependency(jars = NULL, packages = NULL, initializer = NULL,
  catalog = NULL, ...)
}
\arguments{
\item{jars}{Character vector of full paths to JAR files.}

\item{packages}{Character vector of Spark packages names.}

\item{initializer}{Optional callback function called when initializing a connection.}

\item{catalog}{Optional location where extension JAR files can be downloaded for Livy.}

\item{...}{Additional optional arguments.}
}
\value{
An object of type `spark_dependency`
}
\description{
Define a Spark dependency consisting of a set of custom JARs and Spark packages.
}
