% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_disconnect}
\alias{spark_disconnect}
\alias{spark_disconnect_all}
\title{Disconnect from Spark}
\usage{
spark_disconnect(x, terminate = FALSE, ...)

spark_disconnect_all(termiante = FALSE)
}
\arguments{
\item{x}{A \code{\link{spark_connect}} connection or the Spark master}

\item{terminate}{\code{TRUE} to terminate \code{\link{spark_connect}}
launched in \code{service = TRUE} mode.}

\item{...}{Additional arguments}
}
\description{
Disconnect from Spark

Close all existing connections
}
\seealso{
Other Spark connections: \code{\link{spark_connect}}

Other Spark connections: \code{\link{spark_connect}}
}

