% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_disconnect}
\alias{spark_disconnect}
\alias{spark_disconnect_all}
\title{Disconnect from Spark}
\usage{
spark_disconnect(sc)

spark_disconnect_all()
}
\arguments{
\item{sc}{Spark connection provided by \code{\link{spark_connect}}}
}
\description{
Disconnect from Spark

Close all existing connections
}
\seealso{
Other Spark connections: \code{\link{spark_connection_is_open}},
  \code{\link{spark_connect}}, \code{\link{spark_log}},
  \code{\link{spark_web}}

Other Spark connections: \code{\link{spark_connection_is_open}},
  \code{\link{spark_connect}}, \code{\link{spark_log}},
  \code{\link{spark_web}}
}

