% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_disconnect}
\alias{spark_disconnect}
\alias{spark_disconnect_all}
\title{Disconnects from Spark and terminates the running application}
\usage{
spark_disconnect(sc)

spark_disconnect_all()
}
\arguments{
\item{sc}{Spark connection provided by spark_connect}
}
\description{
Disconnects from Spark and terminates the running application

Closes all existing connections. Returns the total of connections closed.
}

