% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_disconnect}
\alias{spark_disconnect}
\alias{spark_disconnect_all}
\title{Disconnect from Spark}
\usage{
spark_disconnect(x, ...)

spark_disconnect_all()
}
\arguments{
\item{x}{A \code{\link{spark_connect}} connection or the Spark master}

\item{...}{Additional arguments}
}
\description{
Disconnect from Spark

Close all existing connections
}
\seealso{
Other Spark connections: \code{\link{spark_connect}}

Other Spark connections: \code{\link{spark_connect}}
}

