% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R, R/install_spark_versions.R
\name{spark_install}
\alias{spark_install}
\alias{spark_install_tar}
\alias{spark_versions}
\title{Provides support to download and install the given Spark version}
\usage{
spark_install(version = NULL, hadoop_version = NULL, reset = FALSE,
  logging = "INFO", verbose = interactive())

spark_install_tar(path)

spark_versions(latest = FALSE)
}
\arguments{
\item{version}{Version of Spark to install. See spark_versions for a list of supported versions}

\item{hadoop_version}{Version of Hadoop to install. See spark_versions_hadoop for a list of supported versions}

\item{reset}{Attempts to reset settings to defaults}

\item{logging}{Logging level to configure install. Supported options: "WARN", "INFO"}

\item{verbose}{Report information as Spark is downloaded / installed?}

\item{path}{Path to TAR file conforming to the pattern spark-###-bin-hadoop### where ###
reference spark and hadoop versions respectevely.}

\item{latest}{TRUE to use the latest download URLs, FALSE to use the package defaults which might be out of date.}
}
\description{
Provides support to download and install the given Spark version

Provides support to install a version of Spark from a given TAR file

Retrieves available versions of Spark
}

