% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R
\name{spark_install}
\alias{spark_install}
\title{Provides support to download and install the given Spark version}
\usage{
spark_install(version = "1.6.0", hadoop_version = "2.6", reset = FALSE,
  logging = "INFO")
}
\arguments{
\item{version}{Version of Spark to install. See spark_versions for a list of supported versions}

\item{hadoop_version}{Version of Hadoop to install. See spark_versions_hadoop for a list of supported versions}

\item{reset}{Attempts to reset settings to defaults}

\item{logging}{Logging level to configure install. Supported options: "WARN", "INFO"}
}
\description{
Provides support to download and install the given Spark version
}

