% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark.R, R/install_spark_versions.R
\name{spark_install}
\alias{spark_available_versions}
\alias{spark_install}
\alias{spark_install_tar}
\alias{spark_installed_versions}
\title{Download and install various versions of Spark}
\usage{
spark_install(version = NULL, hadoop_version = NULL, reset = FALSE,
  logging = "INFO", verbose = interactive())

spark_install_tar(tarfile)

spark_installed_versions()

spark_available_versions()
}
\arguments{
\item{version}{Version of Spark to install. See \code{spark_available_versions} for a list of supported versions}

\item{hadoop_version}{Version of Hadoop to install. See \code{spark_available_versions} for a list of supported versions}

\item{reset}{Attempts to reset settings to defaults}

\item{logging}{Logging level to configure install. Supported options: "WARN", "INFO"}

\item{verbose}{Report information as Spark is downloaded / installed?}

\item{tarfile}{Path to TAR file conforming to the pattern spark-###-bin-hadoop### where ###
reference spark and hadoop versions respectevely.}
}
\description{
Download and install various versions of Spark
}

