% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_invoke}
\alias{spark_invoke}
\title{Executes a method on the given object}
\usage{
spark_invoke(jobj, methodName, ...)
}
\arguments{
\item{jobj}{Reference to a jobj retrieved using spark_invoke.
Can alternately be a Spark connection, in this case it is
converted to the Spark context jobj via the
\code{\link{spark_context}} function.}

\item{methodName}{Name of class method to execute}

\item{...}{Additional parameters that method requires}
}
\description{
Executes a method on the given object
}

