% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_invoke_static}
\alias{spark_invoke_static}
\title{Executes an static method on the given object}
\usage{
spark_invoke_static(sconref, objName, methodName, ...)
}
\arguments{
\item{sconref}{Spark connection reference provided by spark_connect}

\item{objName}{Fully-qualified name to static class}

\item{methodName}{Name of class method to execute}

\item{...}{Additional parameters that method requires}
}
\description{
Executes an static method on the given object
}

