% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_log}
\alias{spark_log}
\alias{spark_log_file}
\title{Retrieves entries from the Spark log}
\usage{
spark_log(sc, n = 100)

spark_log_file(sc)
}
\arguments{
\item{sc}{Spark connection provided by \code{\link{spark_connect}}}

\item{n}{Max number of log entries to retrieve}
}
\value{
Character vector with last \code{n} lines of the Spark log 
  or for \code{spark_log_file} the full path to the log file.
}
\description{
Retrieves entries from the Spark log
}
\seealso{
Other Spark connections: \code{\link{spark_connection_is_open}},
  \code{\link{spark_connect}},
  \code{\link{spark_disconnect}}, \code{\link{spark_web}}
}

