% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_log}
\alias{spark_log}
\alias{spark_log_file}
\title{Retrieves the last n entries in the Spark log}
\usage{
spark_log(sc, n = 100)

spark_log_file(sc)
}
\arguments{
\item{sc}{Spark connection provided by spark_connect}

\item{n}{Max number of log entries to retrieve}
}
\description{
Retrieves the last n entries in the Spark log
}

