% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_log}
\alias{print.spark_log}
\alias{spark_log}
\alias{spark_log_file}
\title{Retrieves the last n entries in the Spark log}
\usage{
spark_log(sc, n = 100)

spark_log_file(sc)

\method{print}{spark_log}(x, ...)
}
\arguments{
\item{sc}{Spark connection provided by spark_connect}

\item{n}{Max number of log entries to retrieve}

\item{x}{Spark connection provided by spark_connect}

\item{...}{Additional parameters}
}
\description{
Retrieves the last n entries in the Spark log

Prints a spark_log object
}

