% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_log}
\alias{spark_log}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sparkapi}{\code{\link[sparkapi]{spark_log}}}
}}
\keyword{internal}

