% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_log}
\alias{print.spark_log}
\alias{spark_log}
\title{Retrieves the last n entries in the Spark log}
\usage{
spark_log(sconref, n = 100)

\method{print}{spark_log}(x, ...)
}
\arguments{
\item{sconref}{Spark connection reference provided by spark_connect}

\item{n}{Max number of log entries to retrieve}

\item{x}{Spark log provided by spark_log}

\item{...}{Additional parameters}
}
\description{
Retrieves the last n entries in the Spark log

Prints a spark_log object
}

