% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_log}
\alias{spark_log}
\title{Retrieves entries from the Spark log}
\usage{
spark_log(sc, n = 100, ...)
}
\arguments{
\item{sc}{\code{spark_connection}}

\item{n}{Max number of log entries to retrieve (pass NULL to retrieve
all lines of the log)}

\item{...}{Unused (reserved for future use)}
}
\value{
Character vector with last \code{n} lines of the Spark log
  or for \code{spark_log_file} the full path to the log file.
}
\description{
Retrieves entries from the Spark log
}

