% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_log}
\alias{spark_log}
\title{View Entries in the Spark Log}
\usage{
spark_log(sc, n = 100, filter = NULL, ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{n}{The max number of log entries to retrieve. Use \code{NULL} to
retrieve all entries within the log.}

\item{filter}{Character string to filter log entries.}

\item{...}{Optional arguments; currently unused.}
}
\description{
View the most recent entries in the Spark log. This can be useful when
inspecting output / errors produced by Spark during the invocation of
various commands.
}
