% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{spark_save_table}
\alias{spark_save_table}
\title{Saves a Spark DataFrame as a Spark table}
\usage{
spark_save_table(x, path, mode = NULL)
}
\arguments{
\item{x}{A Spark DataFrame or dplyr operation}

\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3n://"} and \samp{"file://"} protocols.}

\item{mode}{Specifies the behavior when data or table already exists.}
}
\description{
Saves a Spark DataFrame and as a Spark table.
}
\seealso{
Other Spark serialization routines: \code{\link{spark_load_table}},
  \code{\link{spark_read_csv}},
  \code{\link{spark_read_json}},
  \code{\link{spark_read_parquet}},
  \code{\link{spark_write_csv}},
  \code{\link{spark_write_json}},
  \code{\link{spark_write_parquet}}
}

