% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_version.R
\name{spark_version}
\alias{spark_version}
\title{Get the Spark Version Associated with a Spark Connection}
\usage{
spark_version(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}
}
\value{
The Spark version as a \code{\link{numeric_version}}.
}
\description{
Retrieve the version of Spark associated with a Spark connection.
}
\details{
Suffixes for e.g. preview versions, or snapshotted versions,
are trimmed -- if you require the full Spark version, you can
retrieve it with \code{invoke(spark_context(sc), "version")}.
}
