% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions}
\alias{spark_versions}
\title{Retrieves available versions of Spark}
\usage{
spark_versions(supported = TRUE)
}
\arguments{
\item{supported}{If TRUE (default), returns only the fully supported versions in this package.
Otherwise, retrieves all versions which may be only partially supported.}
}
\description{
Retrieves available versions of Spark
}

