% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions}
\alias{spark_versions}
\title{Returns a data frame of available Spark versions that can be installed.}
\usage{
spark_versions(latest = TRUE)
}
\arguments{
\item{latest}{Check for latest version?}
}
\description{
Returns a data frame of available Spark versions that can be installed.
}
\keyword{internal}
