% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_spark_versions.R
\name{spark_versions}
\alias{spark_versions}
\title{Retrieves a dataframe available Spark versions that van be installed.}
\usage{
spark_versions(latest = TRUE)
}
\arguments{
\item{latest}{Check for latest version?}
}
\description{
Retrieves a dataframe available Spark versions that van be installed.
}
\keyword{internal}
