% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_web}
\alias{spark_web}
\title{Open the Spark web interface}
\usage{
spark_web(sc)
}
\arguments{
\item{sc}{Spark connection provided by \code{\link{spark_connect}}}
}
\description{
Open the Spark web interface
}
\seealso{
Other Spark connections: \code{\link{spark_connection_is_open}},
  \code{\link{spark_connect}},
  \code{\link{spark_disconnect}}, \code{\link{spark_log}}
}

