% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_spark.R
\name{spark_web}
\alias{spark_web}
\title{Opens the Spark web interface}
\usage{
spark_web(sc)
}
\arguments{
\item{sc}{Spark connection provided by spark_connect}
}
\description{
Opens the Spark web interface
}

