% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_web}
\alias{spark_web}
\title{Open the Spark web interface}
\usage{
spark_web(sc, ...)
}
\arguments{
\item{sc}{\code{spark_connection}}

\item{...}{Unused (reserved for future use)}
}
\description{
Open the Spark web interface
}

