% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_connection.R
\name{spark_web}
\alias{spark_web}
\title{Open the Spark web interface}
\usage{
spark_web(sc, ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Open the Spark web interface
}
