% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark_data.R
\name{spark_write_json}
\alias{spark_write_json}
\title{Writes a dplyr operation result as a JSON file}
\usage{
spark_write_json(x, path)
}
\arguments{
\item{x}{A dplyr operation, for instance, `tbls(db, "flights")`}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports: "hdfs://" or "s3n://"}
}
\description{
Writes a dplyr operation result as a JSON file
}

