% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{spark_write_parquet}
\alias{spark_write_parquet}
\title{Write a Spark DataFrame to a Parquet file}
\usage{
spark_write_parquet(x, path, mode = NULL, options = list())
}
\arguments{
\item{x}{A Spark DataFrame or dplyr operation}

\item{path}{The path to the file. Needs to be accessible from the cluster. Supports: "hdfs://" or "s3n://"}

\item{mode}{Specifies the behavior when data or table already exists.}

\item{options}{A list of strings with additional options. See \url{http://spark.apache.org/docs/latest/sql-programming-guide.html#configuration}.}
}
\description{
Write a Spark DataFrame to a Parquet file
}
\seealso{
Other reading and writing data: \code{\link{spark_read_csv}},
  \code{\link{spark_read_json}},
  \code{\link{spark_read_parquet}},
  \code{\link{spark_write_csv}},
  \code{\link{spark_write_json}}
}

