% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_spark.R
\name{src_databases}
\alias{src_databases}
\title{Show database list}
\usage{
src_databases(sc, ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Show database list
}
