% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_data.R
\name{stream_read_memory}
\alias{stream_read_memory}
\title{Read a Spark DataFrame Memory Stream}
\usage{
stream_read_memory(sc, source = NULL, name = NULL, header = TRUE,
  options = list(), ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{source}{The name of the memory stream to read from.}

\item{name}{The name to assign to the newly generated stream.}

\item{header}{Boolean; should the first row of data be used as a header?
Defaults to \code{TRUE}.}

\item{options}{A list of strings with additional options.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Read a memory data stream into a Spark DataFrame.
}
\seealso{
Other Spark stream serialization: \code{\link{stream_read_csv}},
  \code{\link{stream_write_csv}},
  \code{\link{stream_write_memory}}
}
