% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_data.R
\name{stream_write_csv}
\alias{stream_write_csv}
\title{Write a Spark DataFrame to a CSV stream}
\usage{
stream_write_csv(x, path)
}
\arguments{
\item{x}{A Spark DataFrame or dplyr operation}

\item{path}{The path to the file. Needs to be accessible from the cluster.
Supports the \samp{"hdfs://"}, \samp{"s3a://"} and \samp{"file://"} protocols.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Write a Spark DataFrame to a tabular (typically, comma-separated) stream
}
\seealso{
Other Spark stream serialization: \code{\link{stream_read_csv}}
}
