% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_data.R
\name{stream_write_memory}
\alias{stream_write_memory}
\title{Write Memory Stream}
\usage{
stream_write_memory(
  x,
  name = random_string("sparklyr_tmp_"),
  mode = c("append", "complete", "update"),
  trigger = stream_trigger_interval(),
  checkpoint = file.path("checkpoints", name, random_string("")),
  options = list(),
  partition_by = NULL,
  ...
)
}
\arguments{
\item{x}{A Spark DataFrame or dplyr operation}

\item{name}{The name to assign to the newly generated stream.}

\item{mode}{Specifies how data is written to a streaming sink. Valid values are
\code{"append"}, \code{"complete"} or \code{"update"}.}

\item{trigger}{The trigger for the stream query, defaults to micro-batches
running every 5 seconds. See \code{\link{stream_trigger_interval}} and
  \code{\link{stream_trigger_continuous}}.}

\item{checkpoint}{The location where the system will write all the checkpoint
information to guarantee end-to-end fault-tolerance.}

\item{options}{A list of strings with additional options.}

\item{partition_by}{Partitions the output by the given list of columns.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Writes a Spark dataframe stream into a memory stream.
}
\seealso{
Other Spark stream serialization: 
\code{\link{stream_write_csv}()}
}
\concept{Spark stream serialization}
