% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{tbl_cache}
\alias{tbl_cache}
\title{Cache a Spark Table}
\usage{
tbl_cache(sc, name, force = TRUE)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{The table name.}

\item{force}{Force the data to be loaded into memory? This is accomplished
by calling the \code{count} API on the associated Spark DataFrame.}
}
\description{
Force a Spark table with name \code{name} to be loaded into memory.
Operations on cached tables should normally (although not always)
be more performant than the same operation performed on an uncached
table.
}
\seealso{
Other dplyr: \code{\link{copy_to}},
  \code{\link{tbl_uncache}}
}

