% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{tbl_cache}
\alias{tbl_cache}
\title{Loads a table into memory}
\usage{
tbl_cache(con, name, force = TRUE)
}
\arguments{
\item{con}{Connection to dplyr source}

\item{name}{Name of the destination table}

\item{force}{Forces data to be loaded in memory by executing a count(*) over the table}
}
\description{
Loads a table into memory
}

