% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{tbl_cache}
\alias{tbl_cache}
\title{Load a table into memory}
\usage{
tbl_cache(sc, name, force = TRUE)
}
\arguments{
\item{sc}{Spark connection}

\item{name}{Name of the destination table}

\item{force}{Forces data to be loaded in memory by executing a count(*) over the table}
}
\description{
Load a table into memory
}
\seealso{
Other dplyr: \code{\link{copy_to}},
  \code{\link{tbl_uncache}}
}

