% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{tbl_uncache}
\alias{tbl_uncache}
\title{Unloads table from memory}
\usage{
tbl_uncache(con, name)
}
\arguments{
\item{con}{Connection to dplyr source}

\item{name}{Name of the destination table}
}
\description{
Unloads table from memory
}

