% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_spark.R
\name{tbl_uncache}
\alias{tbl_uncache}
\title{Unload table from memory}
\usage{
tbl_uncache(sc, name)
}
\arguments{
\item{sc}{Spark connection}

\item{name}{Name of the destination table}
}
\description{
Unload table from memory
}
\seealso{
Other dplyr: \code{\link{copy_to}}, \code{\link{tbl_cache}}
}

