% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_spark.R
\name{tbl_uncache}
\alias{tbl_uncache}
\title{Uncache a Spark Table}
\usage{
tbl_uncache(sc, name)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{The table name.}
}
\description{
Force a Spark table with name \code{name} to be unloaded from memory.
}

